Class "EnvModules"

The function module enables use of the Environment Modules system (http://modules.sourceforge.net/) from within the R environment. By default the user's login shell environment (ie. bash -l) will be used to initialize the current session. The module function can also; load or unload specific software, list all the loaded software within the current session, and list all the applications available for loading from the module system. Lastly, the module function can remove all loaded software from the current session.

Objects from the Class

Objects can be created by calls of the form new("EnvModules", ...).

Slots

available_modules:

Object of class "list" ~~

loaded_modules:

Object of class "list" ~~

default_modules:

Object of class "list" ~~

modulecmd:

Object of class "character" ~~

Methods

[

signature(x = "EnvModules"): ...

[[

signature(x = "EnvModules", i = "ANY", j = "missing"): ...

[[<-

signature(x = "EnvModules"): ...

$

signature(x = "EnvModules"): ...

available_modules

signature(x = "EnvModules"): ...

coerce

signature(from = "EnvModules", to = "list"): ...

coerce

signature(from = "list", to = "EnvModules"): ...

default_modules

signature(x = "EnvModules"): ...

EnvModules

signature(x = "EnvModules"): ...

loaded_modules

signature(x = "EnvModules"): ...

modulecmd

signature(x = "EnvModules"): ...

names

signature(x = "EnvModules"): ...

show

signature(object = "EnvModules"): ...

Author

Jordan Hayes and Daniela Cassol

Examples

showClass("EnvModules")

#> Class "EnvModules" [package "systemPipeR"]
#> 
#> Slots:
#>                                                                               
#> Name:  available_modules    loaded_modules   default_modules         modulecmd
#> Class:              list              list              list         character