Class "INTERSECTset"

Container for storing standard intersect results created by the overLapper function. The setlist slot stores the original label sets as vectors in a list; intersectmatrix organizes the label sets in a present-absent matrix; complexitylevels represents the number of comparisons considered for each comparison set as vector of integers; and intersectlist contains the standard intersect vectors.

Objects from the Class

Objects can be created by calls of the form new("INTERSECTset", ...).

Slots

setlist:

Object of class "list": list of vectors

intersectmatrix:

Object of class "matrix": binary matrix

complexitylevels:

Object of class "integer": vector of integers

intersectlist:

Object of class "list": list of vectors

Methods

as.list

signature(x = "INTERSECTset"): coerces INTERSECTset to list

coerce

signature(from = "list", to = "INTERSECTset"): as(list, "INTERSECTset")

complexitylevels

signature(x = "INTERSECTset"): extracts data from complexitylevels slot

intersectlist

signature(x = "INTERSECTset"): extracts data from intersectlist slot

intersectmatrix

signature(x = "INTERSECTset"): extracts data from intersectmatrix slot

length

signature(x = "INTERSECTset"): returns number of original label sets

names

signature(x = "INTERSECTset"): extracts slot names

setlist

signature(x = "INTERSECTset"): extracts data from setlist slot

show

signature(object = "INTERSECTset"): summary view of INTERSECTset objects

Author

Thomas Girke

See also

Examples

showClass("INTERSECTset")

#> Class "INTERSECTset" [package "systemPipeR"]
#> 
#> Slots:
#>                                                                           
#> Name:           setlist  intersectmatrix complexitylevels    intersectlist
#> Class:             list           matrix          integer             list

## Sample data
setlist <- list(A=sample(letters, 18), B=sample(letters, 16),
                C=sample(letters, 20), D=sample(letters, 22),
                E=sample(letters, 18), F=sample(letters, 22))

## Create VENNset
interset <- overLapper(setlist[1:5], type="intersects")
class(interset)

#> [1] "INTERSECTset"
#> attr(,"package")
#> [1] "systemPipeR"

## Accessor methods for VENNset/INTERSECTset objects
names(interset)

#> [1] "setlist"          "intersectmatrix"  "complexitylevels" "intersectlist"   

setlist(interset)

#> $A
#>  [1] "m" "w" "l" "e" "f" "o" "u" "d" "v" "i" "y" "p" "h" "z" "b" "g" "t" "j"
#> 
#> $B
#>  [1] "l" "b" "f" "d" "p" "y" "c" "k" "g" "x" "i" "q" "j" "e" "u" "a"
#> 
#> $C
#>  [1] "t" "x" "v" "a" "f" "b" "s" "n" "w" "o" "u" "e" "h" "r" "l" "c" "j" "k" "p"
#> [20] "g"
#> 
#> $D
#>  [1] "j" "c" "f" "d" "z" "a" "g" "x" "b" "h" "u" "l" "p" "t" "o" "s" "i" "r" "k"
#> [20] "m" "w" "v"
#> 
#> $E
#>  [1] "v" "n" "e" "g" "s" "l" "q" "x" "t" "m" "f" "a" "o" "r" "b" "z" "w" "y"
#> 

intersectmatrix(interset)

#>   A B C D E
#> a 0 1 1 1 1
#> b 1 1 1 1 1
#> c 0 1 1 1 0
#> d 1 1 0 1 0
#> e 1 1 1 0 1
#> f 1 1 1 1 1
#> g 1 1 1 1 1
#> h 1 0 1 1 0
#> i 1 1 0 1 0
#> j 1 1 1 1 0
#> k 0 1 1 1 0
#> l 1 1 1 1 1
#> m 1 0 0 1 1
#> n 0 0 1 0 1
#> o 1 0 1 1 1
#> p 1 1 1 1 0
#> q 0 1 0 0 1
#> r 0 0 1 1 1
#> s 0 0 1 1 1
#> t 1 0 1 1 1
#> u 1 1 1 1 0
#> v 1 0 1 1 1
#> w 1 0 1 1 1
#> x 0 1 1 1 1
#> y 1 1 0 0 1
#> z 1 0 0 1 1

complexitylevels(interset)

#>  [1] 1 1 1 1 1 2 2 2 2 2 2 2 2 2 2 3 3 3 3 3 3 3 3 3 3 4 4 4 4 4 5

intersectlist(interset)

#> $A
#>  [1] "b" "d" "e" "f" "g" "h" "i" "j" "l" "m" "o" "p" "t" "u" "v" "w" "y" "z"
#> 
#> $B
#>  [1] "a" "b" "c" "d" "e" "f" "g" "i" "j" "k" "l" "p" "q" "u" "x" "y"
#> 
#> $C
#>  [1] "a" "b" "c" "e" "f" "g" "h" "j" "k" "l" "n" "o" "p" "r" "s" "t" "u" "v" "w"
#> [20] "x"
#> 
#> $D
#>  [1] "a" "b" "c" "d" "f" "g" "h" "i" "j" "k" "l" "m" "o" "p" "r" "s" "t" "u" "v"
#> [20] "w" "x" "z"
#> 
#> $E
#>  [1] "a" "b" "e" "f" "g" "l" "m" "n" "o" "q" "r" "s" "t" "v" "w" "x" "y" "z"
#> 
#> $A_B
#>  [1] "b" "d" "e" "f" "g" "i" "j" "l" "p" "u" "y"
#> 
#> $A_C
#>  [1] "b" "e" "f" "g" "h" "j" "l" "o" "p" "t" "u" "v" "w"
#> 
#> $A_D
#>  [1] "b" "d" "f" "g" "h" "i" "j" "l" "m" "o" "p" "t" "u" "v" "w" "z"
#> 
#> $A_E
#>  [1] "b" "e" "f" "g" "l" "m" "o" "t" "v" "w" "y" "z"
#> 
#> $B_C
#>  [1] "a" "b" "c" "e" "f" "g" "j" "k" "l" "p" "u" "x"
#> 
#> $B_D
#>  [1] "a" "b" "c" "d" "f" "g" "i" "j" "k" "l" "p" "u" "x"
#> 
#> $B_E
#> [1] "a" "b" "e" "f" "g" "l" "q" "x" "y"
#> 
#> $C_D
#>  [1] "a" "b" "c" "f" "g" "h" "j" "k" "l" "o" "p" "r" "s" "t" "u" "v" "w" "x"
#> 
#> $C_E
#>  [1] "a" "b" "e" "f" "g" "l" "n" "o" "r" "s" "t" "v" "w" "x"
#> 
#> $D_E
#>  [1] "a" "b" "f" "g" "l" "m" "o" "r" "s" "t" "v" "w" "x" "z"
#> 
#> $A_B_C
#> [1] "b" "e" "f" "g" "j" "l" "p" "u"
#> 
#> $A_B_D
#> [1] "b" "d" "f" "g" "i" "j" "l" "p" "u"
#> 
#> $A_B_E
#> [1] "b" "e" "f" "g" "l" "y"
#> 
#> $A_C_D
#>  [1] "b" "f" "g" "h" "j" "l" "o" "p" "t" "u" "v" "w"
#> 
#> $A_C_E
#> [1] "b" "e" "f" "g" "l" "o" "t" "v" "w"
#> 
#> $A_D_E
#>  [1] "b" "f" "g" "l" "m" "o" "t" "v" "w" "z"
#> 
#> $B_C_D
#>  [1] "a" "b" "c" "f" "g" "j" "k" "l" "p" "u" "x"
#> 
#> $B_C_E
#> [1] "a" "b" "e" "f" "g" "l" "x"
#> 
#> $B_D_E
#> [1] "a" "b" "f" "g" "l" "x"
#> 
#> $C_D_E
#>  [1] "a" "b" "f" "g" "l" "o" "r" "s" "t" "v" "w" "x"
#> 
#> $A_B_C_D
#> [1] "b" "f" "g" "j" "l" "p" "u"
#> 
#> $A_B_C_E
#> [1] "b" "e" "f" "g" "l"
#> 
#> $A_B_D_E
#> [1] "b" "f" "g" "l"
#> 
#> $A_C_D_E
#> [1] "b" "f" "g" "l" "o" "t" "v" "w"
#> 
#> $B_C_D_E
#> [1] "a" "b" "f" "g" "l" "x"
#> 
#> $A_B_C_D_E
#> [1] "b" "f" "g" "l"
#> 

## Coerce VENNset/INTERSECTset object to list
as.list(interset)

#> $setlist
#> $setlist$A
#>  [1] "m" "w" "l" "e" "f" "o" "u" "d" "v" "i" "y" "p" "h" "z" "b" "g" "t" "j"
#> 
#> $setlist$B
#>  [1] "l" "b" "f" "d" "p" "y" "c" "k" "g" "x" "i" "q" "j" "e" "u" "a"
#> 
#> $setlist$C
#>  [1] "t" "x" "v" "a" "f" "b" "s" "n" "w" "o" "u" "e" "h" "r" "l" "c" "j" "k" "p"
#> [20] "g"
#> 
#> $setlist$D
#>  [1] "j" "c" "f" "d" "z" "a" "g" "x" "b" "h" "u" "l" "p" "t" "o" "s" "i" "r" "k"
#> [20] "m" "w" "v"
#> 
#> $setlist$E
#>  [1] "v" "n" "e" "g" "s" "l" "q" "x" "t" "m" "f" "a" "o" "r" "b" "z" "w" "y"
#> 
#> 
#> $intersectmatrix
#>   A B C D E
#> a 0 1 1 1 1
#> b 1 1 1 1 1
#> c 0 1 1 1 0
#> d 1 1 0 1 0
#> e 1 1 1 0 1
#> f 1 1 1 1 1
#> g 1 1 1 1 1
#> h 1 0 1 1 0
#> i 1 1 0 1 0
#> j 1 1 1 1 0
#> k 0 1 1 1 0
#> l 1 1 1 1 1
#> m 1 0 0 1 1
#> n 0 0 1 0 1
#> o 1 0 1 1 1
#> p 1 1 1 1 0
#> q 0 1 0 0 1
#> r 0 0 1 1 1
#> s 0 0 1 1 1
#> t 1 0 1 1 1
#> u 1 1 1 1 0
#> v 1 0 1 1 1
#> w 1 0 1 1 1
#> x 0 1 1 1 1
#> y 1 1 0 0 1
#> z 1 0 0 1 1
#> 
#> $complexitylevels
#>  [1] 1 1 1 1 1 2 2 2 2 2 2 2 2 2 2 3 3 3 3 3 3 3 3 3 3 4 4 4 4 4 5
#> 
#> $intersectlist
#> $intersectlist$A
#>  [1] "b" "d" "e" "f" "g" "h" "i" "j" "l" "m" "o" "p" "t" "u" "v" "w" "y" "z"
#> 
#> $intersectlist$B
#>  [1] "a" "b" "c" "d" "e" "f" "g" "i" "j" "k" "l" "p" "q" "u" "x" "y"
#> 
#> $intersectlist$C
#>  [1] "a" "b" "c" "e" "f" "g" "h" "j" "k" "l" "n" "o" "p" "r" "s" "t" "u" "v" "w"
#> [20] "x"
#> 
#> $intersectlist$D
#>  [1] "a" "b" "c" "d" "f" "g" "h" "i" "j" "k" "l" "m" "o" "p" "r" "s" "t" "u" "v"
#> [20] "w" "x" "z"
#> 
#> $intersectlist$E
#>  [1] "a" "b" "e" "f" "g" "l" "m" "n" "o" "q" "r" "s" "t" "v" "w" "x" "y" "z"
#> 
#> $intersectlist$A_B
#>  [1] "b" "d" "e" "f" "g" "i" "j" "l" "p" "u" "y"
#> 
#> $intersectlist$A_C
#>  [1] "b" "e" "f" "g" "h" "j" "l" "o" "p" "t" "u" "v" "w"
#> 
#> $intersectlist$A_D
#>  [1] "b" "d" "f" "g" "h" "i" "j" "l" "m" "o" "p" "t" "u" "v" "w" "z"
#> 
#> $intersectlist$A_E
#>  [1] "b" "e" "f" "g" "l" "m" "o" "t" "v" "w" "y" "z"
#> 
#> $intersectlist$B_C
#>  [1] "a" "b" "c" "e" "f" "g" "j" "k" "l" "p" "u" "x"
#> 
#> $intersectlist$B_D
#>  [1] "a" "b" "c" "d" "f" "g" "i" "j" "k" "l" "p" "u" "x"
#> 
#> $intersectlist$B_E
#> [1] "a" "b" "e" "f" "g" "l" "q" "x" "y"
#> 
#> $intersectlist$C_D
#>  [1] "a" "b" "c" "f" "g" "h" "j" "k" "l" "o" "p" "r" "s" "t" "u" "v" "w" "x"
#> 
#> $intersectlist$C_E
#>  [1] "a" "b" "e" "f" "g" "l" "n" "o" "r" "s" "t" "v" "w" "x"
#> 
#> $intersectlist$D_E
#>  [1] "a" "b" "f" "g" "l" "m" "o" "r" "s" "t" "v" "w" "x" "z"
#> 
#> $intersectlist$A_B_C
#> [1] "b" "e" "f" "g" "j" "l" "p" "u"
#> 
#> $intersectlist$A_B_D
#> [1] "b" "d" "f" "g" "i" "j" "l" "p" "u"
#> 
#> $intersectlist$A_B_E
#> [1] "b" "e" "f" "g" "l" "y"
#> 
#> $intersectlist$A_C_D
#>  [1] "b" "f" "g" "h" "j" "l" "o" "p" "t" "u" "v" "w"
#> 
#> $intersectlist$A_C_E
#> [1] "b" "e" "f" "g" "l" "o" "t" "v" "w"
#> 
#> $intersectlist$A_D_E
#>  [1] "b" "f" "g" "l" "m" "o" "t" "v" "w" "z"
#> 
#> $intersectlist$B_C_D
#>  [1] "a" "b" "c" "f" "g" "j" "k" "l" "p" "u" "x"
#> 
#> $intersectlist$B_C_E
#> [1] "a" "b" "e" "f" "g" "l" "x"
#> 
#> $intersectlist$B_D_E
#> [1] "a" "b" "f" "g" "l" "x"
#> 
#> $intersectlist$C_D_E
#>  [1] "a" "b" "f" "g" "l" "o" "r" "s" "t" "v" "w" "x"
#> 
#> $intersectlist$A_B_C_D
#> [1] "b" "f" "g" "j" "l" "p" "u"
#> 
#> $intersectlist$A_B_C_E
#> [1] "b" "e" "f" "g" "l"
#> 
#> $intersectlist$A_B_D_E
#> [1] "b" "f" "g" "l"
#> 
#> $intersectlist$A_C_D_E
#> [1] "b" "f" "g" "l" "o" "t" "v" "w"
#> 
#> $intersectlist$B_C_D_E
#> [1] "a" "b" "f" "g" "l" "x"
#> 
#> $intersectlist$A_B_C_D_E
#> [1] "b" "f" "g" "l"
#> 
#>