Class "LineWise"

LineWise class stores all the R chunk from a RMarkdown file.

Objects from the Class

Objects can be created by calls of the form new("LineWise", ...).

Slots

codeLine:

Object of class "character" storing data from Rc chunk.

codeChunkStart:

Object of class "interger" storing Start line from the rmdPath file.

rmdPath:

Object of class "character" storing path from the rmdPath file.

stepName:

Object of class "character" storing Step Name.

dependency:

Object of class "character" storing dependency steps.

Methods

[

signature(x = "LineWise"): subsetting of class with bracket operator

[[

signature(x = "LineWise", i = "ANY", j = "missing"): subsetting of class with bracket operator

[[<-

signature(x = "LineWise"): replacement method for LineWise class

$

signature(x = "LineWise"): extracting slots elements by name

codeLine

signature(x = "LineWise"): ...

coerce

signature(from = "LineWise", to = "list"): as(LineWise, "list")

coerce

signature(from = "list", to = "LineWise"): as(list, "LineWise")

length

signature(x = "LineWise"): extracts number of R code lines

linewise

signature(x = "LineWise"): Coerce back to list as(LineWise, "list")

names

signature(x = "LineWise"): extracts slot names

rmdPath

signature(x = "LineWise"): extracts path of the Rmarkdown files

show

signature(object = "LineWise"): summary view of LineWise objects

appendCodeLine<-

signature(x = "LineWise"): replacement method for append a code line

replaceCodeLine<-

signature(x = "LineWise"): replacement method for replace a code line

Author

Daniela Cassol

See also

Examples

showClass("LineWise")

#> Class "LineWise" [package "systemPipeR"]
#> 
#> Slots:
#>                                                                   
#> Name:        codeLine codeChunkStart        rmdPath       stepName
#> Class:     expression        integer      character      character
#>                                     
#> Name:      dependency         status
#> Class:           list           list