Workflow Project Initiation

Function to construct SYSargsList workflow control environment (S4 object). This function creates and checks the directory structure. If the expected directories are not available, it is possible to create those. The project directory structure expected is:

SPRproject/

data/

param/

results/

The project working directory default structure of the project directories can be modified here.

SPRproject(projPath = getwd(), data = "data", param = "param", results = "results",
                      logs.dir= ".SPRproject", sys.file="SYSargsList.yml", 
                      envir = new.env(), 
                      restart = FALSE, resume=FALSE, 
                      load.envir = FALSE,
                      overwrite = FALSE, silent = FALSE)

Arguments

projPath
a character vector of a full project path name. Default is the current path.

data
a character vector of a data directory name. Default is data. This subdirectory in the project stores all the raw data, reference, and annotation files.

param
a character vector of a param directory name. Default is param. This subdirectory in the project stores all the parameter and configuration files.

results
a character vector of a results directory name. Default is results. This subdirectory in the project stores all the analysis results, including but not limited: alignment, variant, and peak files (BAM, VCF, BED); tabular result files; and image/plot files.

logs.dir
a character vector of a logs directory name. Default is .SPRproject.

sys.file
name and path of the SYSargsList file which will store all the project configuration information. Default is .SPRproject/SYSargsList.yml.

envir
the environment in which expr is to be evaluated. Default will create a new.env(). Another option is to use globalenv().

restart
if set to TRUE, existing SYSargsList save in the directory will be used to restart the workflow. This option requires overwrite=TRUE. All the log will be deleted.

resume
if set to TRUE, existing SYSargsList save in the directory will be used to resume the workflow.

load.envir
after running the workflow, the environment can be saved on disk. This argument allows to load the environment and recover all the objects. This argument can be set as TRUE when restart = TRUE.

overwrite
if set to TRUE, existing logs files or directory of same name will be overwritten.

silent
if set to TRUE, all messages returned by the function will be suppressed.

Details

If an SYSargsList instance was created before or independent of the project initialization, it is possible to append this instance after the project is created. Please see check appendStep<- function.

Value

SPRproject will return a SYSargsList object.

Author

Daniela Cassol

See also

Examples

sal <- SPRproject(projPath = tempdir())

#> Creating directory: /tmp/RtmpC6dq7I/data 
#> Creating directory: /tmp/RtmpC6dq7I/param 
#> Creating directory: /tmp/RtmpC6dq7I/results 
#> Creating directory '/tmp/RtmpC6dq7I/.SPRproject'
#> Creating file '/tmp/RtmpC6dq7I/.SPRproject/SYSargsList.yml'

#> Your current working directory is different from the directory chosen for the Project Workflow.
#> For accurate location of the files and running the Workflow, please set the working directory to 
#> 'setwd(/tmp/RtmpC6dq7I)'

sal

#> Instance of 'SYSargsList': 
#>  No workflow steps added