Class "SYSargs"

S4 class container for storing parameters of command-line- or R-based software. SYSargs instances are constructed by the systemArgs function from two simple tabular files: a targets file and a param file. The latter is optional for workflow steps lacking command-line software. Typically, a SYSargs instance stores all sample-level inputs as well as the paths to the corresponding outputs generated by command-line- or R-based software generating sample-level output files. Each sample level input/outfile operation uses its own SYSargs instance. The outpaths of SYSargs usually define the sample inputs for the next SYSargs instance. This connectivity is achieved by writing the outpaths with the writeTargetsout function to a new targets file that serves as input to the next systemArgs call. By chaining several SYSargs steps together one can construct complex workflows involving many sample-level input/output file operations with any combination of command-line or R-based software.

Objects from the Class

Objects can be created by calls of the form new("SYSargs", ...).

Slots

targetsin:

Object of class "data.frame" storing tabular data from targets input file

targetsout:

Object of class "data.frame" storing tabular data from targets output file

targetsheader:

Object of class "character" storing header/comment lines of targets file

modules:

Object of class "character" storing software versions from module system

software:

Object of class "character" name of executable of command-line software

cores:

Object of class "numeric" number of CPU cores to use

other:

Object of class "character" additional arguments

reference:

Object of class "character" path to reference genome file

results:

Object of class "character" path to results directory

infile1:

Object of class "character" paths to first FASTQ file

infile2:

Object of class "character" paths to second FASTQ file if data is PE

outfile1:

Object of class "character" paths to output files generated by command-line software

sysargs:

Object of class "character" full commands used to execute external software

outpaths:

Object of class "character" paths to final outputs including postprocessing by Rsamtools

Methods

SampleName

signature(x = "SYSargs"): extracts sample names

[

signature(x = "SYSargs"): subsetting of class with bracket operator

coerce

signature(from = "list", to = "SYSargs"): as(list, "SYSargs")

cores

signature(x = "SYSargs"): extracts data from cores slot

infile1

signature(x = "SYSargs"): extracts data from infile1 slot

infile2

signature(x = "SYSargs"): extracts data from infile2 slot

modules

signature(x = "SYSargs"): extracts data from modules slot

names

signature(x = "SYSargs"): extracts slot names

length

signature(x = "SYSargs"): extracts number of samples

other

signature(x = "SYSargs"): extracts data from other slot

outfile1

signature(x = "SYSargs"): extracts data from outfile1 slot

outpaths

signature(x = "SYSargs"): extracts data from outpath slot

reference

signature(x = "SYSargs"): extracts data from reference slot

results

signature(x = "SYSargs"): extracts data from results slot

show

signature(object = "SYSargs"): summary view of SYSargs objects

software

signature(x = "SYSargs"): extracts data from software slot

targetsheader

signature(x = "SYSargs"): extracts data from targetsheader slot

targetsin

signature(x = "SYSargs"): extracts data from targetsin slot

targetsout

signature(x = "SYSargs"): extracts data from targetsout slot

Author

Thomas Girke

See also

Examples

showClass("SYSargs")

#> Class "SYSargs" [package "systemPipeR"]
#> 
#> Slots:
#>                                                                             
#> Name:      targetsin    targetsout targetsheader       modules      software
#> Class:    data.frame    data.frame     character     character     character
#>                                                                             
#> Name:          cores         other     reference       results       infile1
#> Class:       numeric     character     character     character     character
#>                                                               
#> Name:        infile2      outfile1       sysargs      outpaths
#> Class:     character     character     character     character

## Construct SYSargs object from param and targets files 
param <- system.file("extdata", "tophat.param", package="systemPipeR")
targets <- system.file("extdata", "targets.txt", package="systemPipeR")
args <- systemArgs(sysma=param, mytargets=targets)

#> Warning: path[1]="./data/SRR446027_1.fastq.gz": No such file or directory

#> Warning: path[2]="./data/SRR446028_1.fastq.gz": No such file or directory

#> Warning: path[3]="./data/SRR446029_1.fastq.gz": No such file or directory

#> Warning: path[4]="./data/SRR446030_1.fastq.gz": No such file or directory

#> Warning: path[5]="./data/SRR446031_1.fastq.gz": No such file or directory

#> Warning: path[6]="./data/SRR446032_1.fastq.gz": No such file or directory

#> Warning: path[7]="./data/SRR446033_1.fastq.gz": No such file or directory

#> Warning: path[8]="./data/SRR446034_1.fastq.gz": No such file or directory

#> Warning: path[9]="./data/SRR446035_1.fastq.gz": No such file or directory

#> Warning: path[10]="./data/SRR446036_1.fastq.gz": No such file or directory

#> Warning: path[11]="./data/SRR446037_1.fastq.gz": No such file or directory

#> Warning: path[12]="./data/SRR446038_1.fastq.gz": No such file or directory

#> Warning: path[13]="./data/SRR446039_1.fastq.gz": No such file or directory

#> Warning: path[14]="./data/SRR446040_1.fastq.gz": No such file or directory

#> Warning: path[15]="./data/SRR446041_1.fastq.gz": No such file or directory

#> Warning: path[16]="./data/SRR446042_1.fastq.gz": No such file or directory

#> Warning: path[17]="./data/SRR446043_1.fastq.gz": No such file or directory

#> Warning: path[18]="./data/SRR446044_1.fastq.gz": No such file or directory

args

#> An instance of 'SYSargs' for running 'tophat' on 18 samples 

names(args); targetsin(args); targetsout(args); targetsheader(args); 

#>  [1] "targetsin"     "targetsout"    "targetsheader" "modules"      
#>  [5] "software"      "cores"         "other"         "reference"    
#>  [9] "results"       "infile1"       "infile2"       "outfile1"     
#> [13] "sysargs"       "outpaths"     

#>                       FileName SampleName Factor SampleLong Experiment
#> 1  ./data/SRR446027_1.fastq.gz        M1A     M1  Mock.1h.A          1
#> 2  ./data/SRR446028_1.fastq.gz        M1B     M1  Mock.1h.B          1
#> 3  ./data/SRR446029_1.fastq.gz        A1A     A1   Avr.1h.A          1
#> 4  ./data/SRR446030_1.fastq.gz        A1B     A1   Avr.1h.B          1
#> 5  ./data/SRR446031_1.fastq.gz        V1A     V1   Vir.1h.A          1
#> 6  ./data/SRR446032_1.fastq.gz        V1B     V1   Vir.1h.B          1
#> 7  ./data/SRR446033_1.fastq.gz        M6A     M6  Mock.6h.A          1
#> 8  ./data/SRR446034_1.fastq.gz        M6B     M6  Mock.6h.B          1
#> 9  ./data/SRR446035_1.fastq.gz        A6A     A6   Avr.6h.A          1
#> 10 ./data/SRR446036_1.fastq.gz        A6B     A6   Avr.6h.B          1
#> 11 ./data/SRR446037_1.fastq.gz        V6A     V6   Vir.6h.A          1
#> 12 ./data/SRR446038_1.fastq.gz        V6B     V6   Vir.6h.B          1
#> 13 ./data/SRR446039_1.fastq.gz       M12A    M12 Mock.12h.A          1
#> 14 ./data/SRR446040_1.fastq.gz       M12B    M12 Mock.12h.B          1
#> 15 ./data/SRR446041_1.fastq.gz       A12A    A12  Avr.12h.A          1
#> 16 ./data/SRR446042_1.fastq.gz       A12B    A12  Avr.12h.B          1
#> 17 ./data/SRR446043_1.fastq.gz       V12A    V12  Vir.12h.A          1
#> 18 ./data/SRR446044_1.fastq.gz       V12B    V12  Vir.12h.B          1
#>           Date
#> 1  23-Mar-2012
#> 2  23-Mar-2012
#> 3  23-Mar-2012
#> 4  23-Mar-2012
#> 5  23-Mar-2012
#> 6  23-Mar-2012
#> 7  23-Mar-2012
#> 8  23-Mar-2012
#> 9  23-Mar-2012
#> 10 23-Mar-2012
#> 11 23-Mar-2012
#> 12 23-Mar-2012
#> 13 23-Mar-2012
#> 14 23-Mar-2012
#> 15 23-Mar-2012
#> 16 23-Mar-2012
#> 17 23-Mar-2012
#> 18 23-Mar-2012

#>                                                                                                          FileName
#> 1  /home/runner/work/systemPipeR/systemPipeR/docs/reference/results/SRR446027_1.fastq.gz.tophat/accepted_hits.bam
#> 2  /home/runner/work/systemPipeR/systemPipeR/docs/reference/results/SRR446028_1.fastq.gz.tophat/accepted_hits.bam
#> 3  /home/runner/work/systemPipeR/systemPipeR/docs/reference/results/SRR446029_1.fastq.gz.tophat/accepted_hits.bam
#> 4  /home/runner/work/systemPipeR/systemPipeR/docs/reference/results/SRR446030_1.fastq.gz.tophat/accepted_hits.bam
#> 5  /home/runner/work/systemPipeR/systemPipeR/docs/reference/results/SRR446031_1.fastq.gz.tophat/accepted_hits.bam
#> 6  /home/runner/work/systemPipeR/systemPipeR/docs/reference/results/SRR446032_1.fastq.gz.tophat/accepted_hits.bam
#> 7  /home/runner/work/systemPipeR/systemPipeR/docs/reference/results/SRR446033_1.fastq.gz.tophat/accepted_hits.bam
#> 8  /home/runner/work/systemPipeR/systemPipeR/docs/reference/results/SRR446034_1.fastq.gz.tophat/accepted_hits.bam
#> 9  /home/runner/work/systemPipeR/systemPipeR/docs/reference/results/SRR446035_1.fastq.gz.tophat/accepted_hits.bam
#> 10 /home/runner/work/systemPipeR/systemPipeR/docs/reference/results/SRR446036_1.fastq.gz.tophat/accepted_hits.bam
#> 11 /home/runner/work/systemPipeR/systemPipeR/docs/reference/results/SRR446037_1.fastq.gz.tophat/accepted_hits.bam
#> 12 /home/runner/work/systemPipeR/systemPipeR/docs/reference/results/SRR446038_1.fastq.gz.tophat/accepted_hits.bam
#> 13 /home/runner/work/systemPipeR/systemPipeR/docs/reference/results/SRR446039_1.fastq.gz.tophat/accepted_hits.bam
#> 14 /home/runner/work/systemPipeR/systemPipeR/docs/reference/results/SRR446040_1.fastq.gz.tophat/accepted_hits.bam
#> 15 /home/runner/work/systemPipeR/systemPipeR/docs/reference/results/SRR446041_1.fastq.gz.tophat/accepted_hits.bam
#> 16 /home/runner/work/systemPipeR/systemPipeR/docs/reference/results/SRR446042_1.fastq.gz.tophat/accepted_hits.bam
#> 17 /home/runner/work/systemPipeR/systemPipeR/docs/reference/results/SRR446043_1.fastq.gz.tophat/accepted_hits.bam
#> 18 /home/runner/work/systemPipeR/systemPipeR/docs/reference/results/SRR446044_1.fastq.gz.tophat/accepted_hits.bam
#>    SampleName Factor SampleLong Experiment        Date
#> 1         M1A     M1  Mock.1h.A          1 23-Mar-2012
#> 2         M1B     M1  Mock.1h.B          1 23-Mar-2012
#> 3         A1A     A1   Avr.1h.A          1 23-Mar-2012
#> 4         A1B     A1   Avr.1h.B          1 23-Mar-2012
#> 5         V1A     V1   Vir.1h.A          1 23-Mar-2012
#> 6         V1B     V1   Vir.1h.B          1 23-Mar-2012
#> 7         M6A     M6  Mock.6h.A          1 23-Mar-2012
#> 8         M6B     M6  Mock.6h.B          1 23-Mar-2012
#> 9         A6A     A6   Avr.6h.A          1 23-Mar-2012
#> 10        A6B     A6   Avr.6h.B          1 23-Mar-2012
#> 11        V6A     V6   Vir.6h.A          1 23-Mar-2012
#> 12        V6B     V6   Vir.6h.B          1 23-Mar-2012
#> 13       M12A    M12 Mock.12h.A          1 23-Mar-2012
#> 14       M12B    M12 Mock.12h.B          1 23-Mar-2012
#> 15       A12A    A12  Avr.12h.A          1 23-Mar-2012
#> 16       A12B    A12  Avr.12h.B          1 23-Mar-2012
#> 17       V12A    V12  Vir.12h.A          1 23-Mar-2012
#> 18       V12B    V12  Vir.12h.B          1 23-Mar-2012

#> [1] "# Project ID: Arabidopsis - Pseudomonas alternative splicing study (SRA: SRP010938; PMID: 24098335)"                                                                              
#> [2] "# The following line(s) allow to specify the contrasts needed for comparative analyses, such as DEG identification. All possible comparisons can be specified with 'CMPset: ALL'."
#> [3] "# <CMP> CMPset1: M1-A1, M1-V1, A1-V1, M6-A6, M6-V6, A6-V6, M12-A12, M12-V12, A12-V12"                                                                                             
#> [4] "# <CMP> CMPset2: ALL"                                                                                                                                                             

software(args); modules(args); cores(args); outpaths(args) 

#> [1] "tophat"

#> [1] "bowtie2/2.2.5" "tophat/2.0.14"

#> [1] 4

#>                                                                                                              M1A 
#> "/home/runner/work/systemPipeR/systemPipeR/docs/reference/results/SRR446027_1.fastq.gz.tophat/accepted_hits.bam" 
#>                                                                                                              M1B 
#> "/home/runner/work/systemPipeR/systemPipeR/docs/reference/results/SRR446028_1.fastq.gz.tophat/accepted_hits.bam" 
#>                                                                                                              A1A 
#> "/home/runner/work/systemPipeR/systemPipeR/docs/reference/results/SRR446029_1.fastq.gz.tophat/accepted_hits.bam" 
#>                                                                                                              A1B 
#> "/home/runner/work/systemPipeR/systemPipeR/docs/reference/results/SRR446030_1.fastq.gz.tophat/accepted_hits.bam" 
#>                                                                                                              V1A 
#> "/home/runner/work/systemPipeR/systemPipeR/docs/reference/results/SRR446031_1.fastq.gz.tophat/accepted_hits.bam" 
#>                                                                                                              V1B 
#> "/home/runner/work/systemPipeR/systemPipeR/docs/reference/results/SRR446032_1.fastq.gz.tophat/accepted_hits.bam" 
#>                                                                                                              M6A 
#> "/home/runner/work/systemPipeR/systemPipeR/docs/reference/results/SRR446033_1.fastq.gz.tophat/accepted_hits.bam" 
#>                                                                                                              M6B 
#> "/home/runner/work/systemPipeR/systemPipeR/docs/reference/results/SRR446034_1.fastq.gz.tophat/accepted_hits.bam" 
#>                                                                                                              A6A 
#> "/home/runner/work/systemPipeR/systemPipeR/docs/reference/results/SRR446035_1.fastq.gz.tophat/accepted_hits.bam" 
#>                                                                                                              A6B 
#> "/home/runner/work/systemPipeR/systemPipeR/docs/reference/results/SRR446036_1.fastq.gz.tophat/accepted_hits.bam" 
#>                                                                                                              V6A 
#> "/home/runner/work/systemPipeR/systemPipeR/docs/reference/results/SRR446037_1.fastq.gz.tophat/accepted_hits.bam" 
#>                                                                                                              V6B 
#> "/home/runner/work/systemPipeR/systemPipeR/docs/reference/results/SRR446038_1.fastq.gz.tophat/accepted_hits.bam" 
#>                                                                                                             M12A 
#> "/home/runner/work/systemPipeR/systemPipeR/docs/reference/results/SRR446039_1.fastq.gz.tophat/accepted_hits.bam" 
#>                                                                                                             M12B 
#> "/home/runner/work/systemPipeR/systemPipeR/docs/reference/results/SRR446040_1.fastq.gz.tophat/accepted_hits.bam" 
#>                                                                                                             A12A 
#> "/home/runner/work/systemPipeR/systemPipeR/docs/reference/results/SRR446041_1.fastq.gz.tophat/accepted_hits.bam" 
#>                                                                                                             A12B 
#> "/home/runner/work/systemPipeR/systemPipeR/docs/reference/results/SRR446042_1.fastq.gz.tophat/accepted_hits.bam" 
#>                                                                                                             V12A 
#> "/home/runner/work/systemPipeR/systemPipeR/docs/reference/results/SRR446043_1.fastq.gz.tophat/accepted_hits.bam" 
#>                                                                                                             V12B 
#> "/home/runner/work/systemPipeR/systemPipeR/docs/reference/results/SRR446044_1.fastq.gz.tophat/accepted_hits.bam" 

sysargs(args); other(args); reference(args); results(args); infile1(args)

#>                                                                                                                                                                                                                                                          M1A 
#> "tophat -p 4 -g 1 --segment-length 25 -i 30 -I 3000 -o /home/runner/work/systemPipeR/systemPipeR/docs/reference/results/SRR446027_1.fastq.gz.tophat /home/runner/work/systemPipeR/systemPipeR/docs/reference/data/tair10.fasta ./data/SRR446027_1.fastq.gz " 
#>                                                                                                                                                                                                                                                          M1B 
#> "tophat -p 4 -g 1 --segment-length 25 -i 30 -I 3000 -o /home/runner/work/systemPipeR/systemPipeR/docs/reference/results/SRR446028_1.fastq.gz.tophat /home/runner/work/systemPipeR/systemPipeR/docs/reference/data/tair10.fasta ./data/SRR446028_1.fastq.gz " 
#>                                                                                                                                                                                                                                                          A1A 
#> "tophat -p 4 -g 1 --segment-length 25 -i 30 -I 3000 -o /home/runner/work/systemPipeR/systemPipeR/docs/reference/results/SRR446029_1.fastq.gz.tophat /home/runner/work/systemPipeR/systemPipeR/docs/reference/data/tair10.fasta ./data/SRR446029_1.fastq.gz " 
#>                                                                                                                                                                                                                                                          A1B 
#> "tophat -p 4 -g 1 --segment-length 25 -i 30 -I 3000 -o /home/runner/work/systemPipeR/systemPipeR/docs/reference/results/SRR446030_1.fastq.gz.tophat /home/runner/work/systemPipeR/systemPipeR/docs/reference/data/tair10.fasta ./data/SRR446030_1.fastq.gz " 
#>                                                                                                                                                                                                                                                          V1A 
#> "tophat -p 4 -g 1 --segment-length 25 -i 30 -I 3000 -o /home/runner/work/systemPipeR/systemPipeR/docs/reference/results/SRR446031_1.fastq.gz.tophat /home/runner/work/systemPipeR/systemPipeR/docs/reference/data/tair10.fasta ./data/SRR446031_1.fastq.gz " 
#>                                                                                                                                                                                                                                                          V1B 
#> "tophat -p 4 -g 1 --segment-length 25 -i 30 -I 3000 -o /home/runner/work/systemPipeR/systemPipeR/docs/reference/results/SRR446032_1.fastq.gz.tophat /home/runner/work/systemPipeR/systemPipeR/docs/reference/data/tair10.fasta ./data/SRR446032_1.fastq.gz " 
#>                                                                                                                                                                                                                                                          M6A 
#> "tophat -p 4 -g 1 --segment-length 25 -i 30 -I 3000 -o /home/runner/work/systemPipeR/systemPipeR/docs/reference/results/SRR446033_1.fastq.gz.tophat /home/runner/work/systemPipeR/systemPipeR/docs/reference/data/tair10.fasta ./data/SRR446033_1.fastq.gz " 
#>                                                                                                                                                                                                                                                          M6B 
#> "tophat -p 4 -g 1 --segment-length 25 -i 30 -I 3000 -o /home/runner/work/systemPipeR/systemPipeR/docs/reference/results/SRR446034_1.fastq.gz.tophat /home/runner/work/systemPipeR/systemPipeR/docs/reference/data/tair10.fasta ./data/SRR446034_1.fastq.gz " 
#>                                                                                                                                                                                                                                                          A6A 
#> "tophat -p 4 -g 1 --segment-length 25 -i 30 -I 3000 -o /home/runner/work/systemPipeR/systemPipeR/docs/reference/results/SRR446035_1.fastq.gz.tophat /home/runner/work/systemPipeR/systemPipeR/docs/reference/data/tair10.fasta ./data/SRR446035_1.fastq.gz " 
#>                                                                                                                                                                                                                                                          A6B 
#> "tophat -p 4 -g 1 --segment-length 25 -i 30 -I 3000 -o /home/runner/work/systemPipeR/systemPipeR/docs/reference/results/SRR446036_1.fastq.gz.tophat /home/runner/work/systemPipeR/systemPipeR/docs/reference/data/tair10.fasta ./data/SRR446036_1.fastq.gz " 
#>                                                                                                                                                                                                                                                          V6A 
#> "tophat -p 4 -g 1 --segment-length 25 -i 30 -I 3000 -o /home/runner/work/systemPipeR/systemPipeR/docs/reference/results/SRR446037_1.fastq.gz.tophat /home/runner/work/systemPipeR/systemPipeR/docs/reference/data/tair10.fasta ./data/SRR446037_1.fastq.gz " 
#>                                                                                                                                                                                                                                                          V6B 
#> "tophat -p 4 -g 1 --segment-length 25 -i 30 -I 3000 -o /home/runner/work/systemPipeR/systemPipeR/docs/reference/results/SRR446038_1.fastq.gz.tophat /home/runner/work/systemPipeR/systemPipeR/docs/reference/data/tair10.fasta ./data/SRR446038_1.fastq.gz " 
#>                                                                                                                                                                                                                                                         M12A 
#> "tophat -p 4 -g 1 --segment-length 25 -i 30 -I 3000 -o /home/runner/work/systemPipeR/systemPipeR/docs/reference/results/SRR446039_1.fastq.gz.tophat /home/runner/work/systemPipeR/systemPipeR/docs/reference/data/tair10.fasta ./data/SRR446039_1.fastq.gz " 
#>                                                                                                                                                                                                                                                         M12B 
#> "tophat -p 4 -g 1 --segment-length 25 -i 30 -I 3000 -o /home/runner/work/systemPipeR/systemPipeR/docs/reference/results/SRR446040_1.fastq.gz.tophat /home/runner/work/systemPipeR/systemPipeR/docs/reference/data/tair10.fasta ./data/SRR446040_1.fastq.gz " 
#>                                                                                                                                                                                                                                                         A12A 
#> "tophat -p 4 -g 1 --segment-length 25 -i 30 -I 3000 -o /home/runner/work/systemPipeR/systemPipeR/docs/reference/results/SRR446041_1.fastq.gz.tophat /home/runner/work/systemPipeR/systemPipeR/docs/reference/data/tair10.fasta ./data/SRR446041_1.fastq.gz " 
#>                                                                                                                                                                                                                                                         A12B 
#> "tophat -p 4 -g 1 --segment-length 25 -i 30 -I 3000 -o /home/runner/work/systemPipeR/systemPipeR/docs/reference/results/SRR446042_1.fastq.gz.tophat /home/runner/work/systemPipeR/systemPipeR/docs/reference/data/tair10.fasta ./data/SRR446042_1.fastq.gz " 
#>                                                                                                                                                                                                                                                         V12A 
#> "tophat -p 4 -g 1 --segment-length 25 -i 30 -I 3000 -o /home/runner/work/systemPipeR/systemPipeR/docs/reference/results/SRR446043_1.fastq.gz.tophat /home/runner/work/systemPipeR/systemPipeR/docs/reference/data/tair10.fasta ./data/SRR446043_1.fastq.gz " 
#>                                                                                                                                                                                                                                                         V12B 
#> "tophat -p 4 -g 1 --segment-length 25 -i 30 -I 3000 -o /home/runner/work/systemPipeR/systemPipeR/docs/reference/results/SRR446044_1.fastq.gz.tophat /home/runner/work/systemPipeR/systemPipeR/docs/reference/data/tair10.fasta ./data/SRR446044_1.fastq.gz " 

#> [1] "-g 1 --segment-length 25 -i 30 -I 3000"

#> [1] "/home/runner/work/systemPipeR/systemPipeR/docs/reference/data/tair10.fasta"

#> [1] "/home/runner/work/systemPipeR/systemPipeR/docs/reference/results/"

#>                           M1A                           M1B 
#> "./data/SRR446027_1.fastq.gz" "./data/SRR446028_1.fastq.gz" 
#>                           A1A                           A1B 
#> "./data/SRR446029_1.fastq.gz" "./data/SRR446030_1.fastq.gz" 
#>                           V1A                           V1B 
#> "./data/SRR446031_1.fastq.gz" "./data/SRR446032_1.fastq.gz" 
#>                           M6A                           M6B 
#> "./data/SRR446033_1.fastq.gz" "./data/SRR446034_1.fastq.gz" 
#>                           A6A                           A6B 
#> "./data/SRR446035_1.fastq.gz" "./data/SRR446036_1.fastq.gz" 
#>                           V6A                           V6B 
#> "./data/SRR446037_1.fastq.gz" "./data/SRR446038_1.fastq.gz" 
#>                          M12A                          M12B 
#> "./data/SRR446039_1.fastq.gz" "./data/SRR446040_1.fastq.gz" 
#>                          A12A                          A12B 
#> "./data/SRR446041_1.fastq.gz" "./data/SRR446042_1.fastq.gz" 
#>                          V12A                          V12B 
#> "./data/SRR446043_1.fastq.gz" "./data/SRR446044_1.fastq.gz" 

infile2(args); outfile1(args); SampleName(args)

#>  M1A  M1B  A1A  A1B  V1A  V1B  M6A  M6B  A6A  A6B  V6A  V6B M12A M12B A12A A12B 
#>   ""   ""   ""   ""   ""   ""   ""   ""   ""   ""   ""   ""   ""   ""   ""   "" 
#> V12A V12B 
#>   ""   "" 

#>                                                                                            M1A 
#> "/home/runner/work/systemPipeR/systemPipeR/docs/reference/results/SRR446027_1.fastq.gz.tophat" 
#>                                                                                            M1B 
#> "/home/runner/work/systemPipeR/systemPipeR/docs/reference/results/SRR446028_1.fastq.gz.tophat" 
#>                                                                                            A1A 
#> "/home/runner/work/systemPipeR/systemPipeR/docs/reference/results/SRR446029_1.fastq.gz.tophat" 
#>                                                                                            A1B 
#> "/home/runner/work/systemPipeR/systemPipeR/docs/reference/results/SRR446030_1.fastq.gz.tophat" 
#>                                                                                            V1A 
#> "/home/runner/work/systemPipeR/systemPipeR/docs/reference/results/SRR446031_1.fastq.gz.tophat" 
#>                                                                                            V1B 
#> "/home/runner/work/systemPipeR/systemPipeR/docs/reference/results/SRR446032_1.fastq.gz.tophat" 
#>                                                                                            M6A 
#> "/home/runner/work/systemPipeR/systemPipeR/docs/reference/results/SRR446033_1.fastq.gz.tophat" 
#>                                                                                            M6B 
#> "/home/runner/work/systemPipeR/systemPipeR/docs/reference/results/SRR446034_1.fastq.gz.tophat" 
#>                                                                                            A6A 
#> "/home/runner/work/systemPipeR/systemPipeR/docs/reference/results/SRR446035_1.fastq.gz.tophat" 
#>                                                                                            A6B 
#> "/home/runner/work/systemPipeR/systemPipeR/docs/reference/results/SRR446036_1.fastq.gz.tophat" 
#>                                                                                            V6A 
#> "/home/runner/work/systemPipeR/systemPipeR/docs/reference/results/SRR446037_1.fastq.gz.tophat" 
#>                                                                                            V6B 
#> "/home/runner/work/systemPipeR/systemPipeR/docs/reference/results/SRR446038_1.fastq.gz.tophat" 
#>                                                                                           M12A 
#> "/home/runner/work/systemPipeR/systemPipeR/docs/reference/results/SRR446039_1.fastq.gz.tophat" 
#>                                                                                           M12B 
#> "/home/runner/work/systemPipeR/systemPipeR/docs/reference/results/SRR446040_1.fastq.gz.tophat" 
#>                                                                                           A12A 
#> "/home/runner/work/systemPipeR/systemPipeR/docs/reference/results/SRR446041_1.fastq.gz.tophat" 
#>                                                                                           A12B 
#> "/home/runner/work/systemPipeR/systemPipeR/docs/reference/results/SRR446042_1.fastq.gz.tophat" 
#>                                                                                           V12A 
#> "/home/runner/work/systemPipeR/systemPipeR/docs/reference/results/SRR446043_1.fastq.gz.tophat" 
#>                                                                                           V12B 
#> "/home/runner/work/systemPipeR/systemPipeR/docs/reference/results/SRR446044_1.fastq.gz.tophat" 

#>  [1] "M1A"  "M1B"  "A1A"  "A1B"  "V1A"  "V1B"  "M6A"  "M6B"  "A6A"  "A6B" 
#> [11] "V6A"  "V6B"  "M12A" "M12B" "A12A" "A12B" "V12A" "V12B"

## Return sample comparisons
readComp(args, format = "vector", delim = "-")

#> $CMPset1
#> [1] "M1-A1"   "M1-V1"   "A1-V1"   "M6-A6"   "M6-V6"   "A6-V6"   "M12-A12"
#> [8] "M12-V12" "A12-V12"
#> 
#> $CMPset2
#>  [1] "M1-A1"   "M1-V1"   "M1-M6"   "M1-A6"   "M1-V6"   "M1-M12"  "M1-A12" 
#>  [8] "M1-V12"  "A1-V1"   "A1-M6"   "A1-A6"   "A1-V6"   "A1-M12"  "A1-A12" 
#> [15] "A1-V12"  "V1-M6"   "V1-A6"   "V1-V6"   "V1-M12"  "V1-A12"  "V1-V12" 
#> [22] "M6-A6"   "M6-V6"   "M6-M12"  "M6-A12"  "M6-V12"  "A6-V6"   "A6-M12" 
#> [29] "A6-A12"  "A6-V12"  "V6-M12"  "V6-A12"  "V6-V12"  "M12-A12" "M12-V12"
#> [36] "A12-V12"
#> 

## The subsetting operator '[' allows to select specific samples
args[1:4]

#> An instance of 'SYSargs' for running 'tophat' on 4 samples 

if (FALSE) {
## Execute SYSargs on single machine
runCommandline(args=args)

## Execute SYSargs on multiple machines  
qsubargs <- getQsubargs(queue="batch", Nnodes="nodes=1", cores=cores(args), memory="mem=10gb", time="walltime=20:00:00")
qsubRun(appfct="runCommandline(args=args)", appargs=args, qsubargs=qsubargs, Nqsubs=1, submitdir="results", package="systemPipeR")

## Write outpaths to new targets file for next SYSargs step
writeTargetsout(x=args, file="default")
}