Deprecated "SYSargs2Pipe" Class

This class and methods are deprecated and should be replaced by SYSargsList class and methods. SYSargs2Pipe class stores a list SYSargs2 objects. Each SYSargs2 objects stores all the information and instructions needed for processing a set of input files with a specific command-line or a series of command-line within a workflow.

Objects from the Class

Objects can be created by calls of the form new("SYSargs2Pipe", ...).

Slots

WF_steps:

Object of class "list" storing all the SYSargs2 objects

track:

Object of class "list" storing all the output files from each SYSargs2 objects

summaryWF:

Object of class "list" storing the summary of all the expected files exists and how many were missing for each SYSargs2 objects

Methods

[

signature(x = "SYSargs2Pipe", i = "ANY", j = "ANY", drop = "ANY"): subsetting of class with bracket operator

[[

signature(x = "SYSargs2Pipe", i = "ANY", j = "ANY"): subsetting of class with bracket operator

[[<-

signature(x = "SYSargs2Pipe", i = "ANY", j = "ANY", value = "ANY"): replacement method for SYSargs2 class

$

signature(x = "SYSargs2Pipe"): extracting slots elements by name

coerce

signature(from = "list", to = "SYSargs2Pipe"): as(list, "SYSargs2Pipe")

coerce

signature(from = "SYSargs2Pipe", to = "list"): as(SYSargs2Pipe, "list")

length

signature(x = "SYSargs2Pipe"): extracts number of SYSargs2 objects

names

signature(x = "SYSargs2Pipe"): extracts slot names

show

signature(object = "SYSargs2Pipe"): summary view of SYSargs2 objects

summaryWF

signature(x = "SYSargs2Pipe"): extract data from targets slot

SYSargs2Pipe_ls

signature(x = "SYSargs2Pipe"): Coerce back to list as(SYSargs2Pipe, "list")

track

signature(x = "SYSargs2Pipe"): extract data from track slot

WF_steps

signature(x = "SYSargs2Pipe"): extract data from WF_steps slot

Author

Daniela Cassol and Thomas Girke

Details

The following methods are deprecated and will be made defunct; use the replacement indicated below:

SYSargsList: help("SYSargsList-class")

See also

Examples

showClass("SYSargs2Pipe")

#> Class "SYSargs2Pipe" [package "systemPipeR"]
#> 
#> Slots:
#>                                     
#> Name:   WF_steps     track summaryWF
#> Class:      list      list      list

if (FALSE) {
## Construct SYSargs2 object from CWl param, CWL input, and targets files 
targets <- system.file("extdata", "targets.txt", package="systemPipeR")
dir_path <- system.file("extdata/cwl/hisat2/hisat2-se", package="systemPipeR")
WF <- loadWorkflow(targets=targets, wf_file="hisat2-mapping-se.cwl", 
                  input_file="hisat2-mapping-se.yml", dir_path=dir_path)
WF <- renderWF(WF, inputvars=c(FileName="_FASTQ_PATH1_", SampleName="_SampleName_"))
WF

## Keep track 
WF_set <- run_track(WF_ls = c(WF))
WF_steps(WF_set)
track(WF_set)
summaryWF(WF_set)[1]
}