Class "SYSargsList"

S4 class container for storing parameters and information of command-line- or R-based workflow.

SYSargsList instances are constructed by the SYSargsList function.

Usage,

## Constructor

SYSargsList(sysargs = NULL, step_name = "default", 
            targets = NULL, wf_file = NULL, input_file = NULL, dir_path = ".", 
            id = "SampleName",
            inputvars = NULL, rm_targets_col = NULL, 
            dir = TRUE, dependency = NA, 
            run_step = "mandatory",
            run_session = "local",
            run_remote_resources = NULL,
            silent = FALSE, 
            projPath = getOption("projPath", getwd()))

## Accessors

stepsWF(x)
statusWF(x)
targetsWF(x)
outfiles(x)
SE(x)
dependency(x)
projectInfo(x)
runInfo(x)

## Methods

baseCommand(x, ...)
SampleName(x, ...)
stepName(x)
targetsheader(x, ...)
yamlinput(x, ...)
cmdlist(x, ...)
codeLine(x, ...)
copyEnvir(x, list = character(), new.env = globalenv(), silent = FALSE)
viewEnvir(x, silent = FALSE)

## Subset Methods

subset(x, ...)
getColumn(x, step, position = c("outfiles", "targetsWF"), column = 1, 
          names = SampleName(x, step))

## Replacement

dependency(x, step, ...) <- value
appendStep(x, after = length(x), ...) <- value
replaceStep(x, step, step_name = "default") <- value
renameStep(x, step, ...) <- value
yamlinput(x, paramName, ...) <- value
appendCodeLine(x, after = length(x), ...) <- value
replaceCodeLine(x, line, ...) <- value
updateColumn(x, step, position = c("outfiles", "targetsWF")) <- value

Arguments

sysargs

SYSargs2 object.

step_name

character with the step index name.

targets

the path to targets file. The targets file can be either a simple tabular or yml file. Also, it is possible to assign NULL to run the pipeline without the 'targets' file. This can be useful for running specific workflows that do not require input files.

wf_file

name and path to CWL parameters file.

input_file

name and path to input parameters file.

dir_path

full path to the directory with the CWL parameters and input files.

id

A column from targets file, which will be used as an id for each one of the samples. It is required to be unique.

inputvars

variables list defined in the input file that matches the column names defined in the targets file.

rm_targets_col

targets file colunms to be removed.

dir

This option allows creating an exclusive results folder for each step in the workflow. All the outfiles and log files for the particular step will be created in the respective folders. Default is dir=TRUE.

dependency

character. Dependency tree, required when appending this step to the workflow. Character name of a previous step in the workflow. Default is NA.

run_step

character. If the step has "mandatory" or "optional" flag for the execution.

run_session

character. If the step has "local" or "remote" flag for the execution.

run_remote_resources

List for reserving for each cluster job sufficient computing resources including memory (Megabyte), number of nodes, CPU cores, walltime (minutes), etc. For more details, one can consult the template file for each queueing/scheduling system.

silent

If set to TRUE, all messages returned by the function will be suppressed.

projPath

a character vector of a full project path name. Default is the current path.

x

An instance of class SYSargsList.

step

character or numeric. Step name or index.

list

a character vector naming objects to be copyied from the enviroment.

new.env

An environment to copy to. Default is globalenv().

position

character. Options are "outfiles" or "targetsWF" slots.

column

character or numeric. Which column will be replaces from the position argument.

names

character vector. Names of the new column.

after

A subscript, after which the values are to be appended.

paramName

character. Input name from yamlinput(x), which value should be replaced.

line

An instance of class LineWise.

value

object containing the values to be replaced to x. Values may be of the same class as the orginal values. For updateColumn<-a dataframe that must have the same rows as the modifying dataframe. However, if there is no column in the original dataframe. This new df will replace the empty one. If there is a non-empty dataframe, any existing columns with the same name as this new df will be replaced. Any columns that do not exist in the original column but this new df will be added to the original data.frame.

...

object containing the values to be replaced to x. Values may be of the same class as the orginal values.

Objects from the Class

Objects can be created by calls of the form new("SYSargsList", ...).

Slots

stepsWF:

Object of class "list" storing all the SYSargs2 objects of the workflow

statusWF:

Object of class "list" storing all the success and failure of each step in the workflow

targetsWF:

Object of class "list" storing all the targets files for each step in the workflow

outfiles:

Object of class "list" storing all the output files in the workflow

SE:

Object of class "list" storing all the SummarizedExperiment objects in the workflow

dependency:

Object of class "list" storing all the dependency steps in the workflow

targets_connection:

Object of class "list" storing all targets files connection in the workflow

projectInfo:

Object of class "list" storing all the projectInfo information of the workflow

runInfo:

Object of class "list" storing all the runInfo information of the workflow

Methods

[

signature(x = "SYSargsList", i = "ANY", j = "ANY", drop = "ANY"): subsetting of class with bracket operator

[[

signature(x = "SYSargsList", i = "ANY", j = "ANY"): subsetting of class with bracket operator

[[<-

signature(x = "SYSargsList", i = "ANY", j = "ANY", value = "ANY"): replacement method for SYSargsList class

$

signature(x = "SYSargsList"): extracting slots elements by name

coerce

signature(from = "list", to = "SYSargsList"): as(list, "SYSargsList")

coerce

signature(from = "SYSargsList", to = "list"): as(SYSargsList, "list")

length

signature(x = "SYSargsList"): extracts number of SYSargs2 objects

names

signature(x = "SYSargsList"): extracts slot names

show

signature(object = "SYSargsList"): summary view of SYSargsList steps

stepsWF

signature(x = "SYSargsList"): extract data from stepsWF slot

statusWF

signature(x = "SYSargsList"): extract data from statusWF slot

dependency

signature(x = "SYSargsList"): extract data from dependency slot

projectInfo

signature(x = "SYSargsList"): extract data from projectInfo slot

runInfo

signature(x = "SYSargsList"): extract data from runInfo slot

targetsWF

signature(x = "SYSargsList"): extract data from targetsWF slot

sysargslist

signature(x = "SYSargsList"): Coerce back to list as(SYSargsList, "list")

SE

signature(x = "SYSargsList"): extract data from SE slot

cmdlist

signature(x = "SYSargsList", sample=NULL): extracts data from cmdlist slot for each step

appendStep<-

signature(x = "SYSargsList"): replacement method for append a step

appendStep<-

signature(x = "SYSargsList"): replacement method for append a step

replaceStep<-

signature(x = "SYSargsList"): replacement method for replace a step

renameStep<-

signature(x = "SYSargsList"): replacement method for rename a step

yamlinput<-

signature(x = "SYSargsList"): replacement method for yamlinput slot for each step

dependency<-

signature(x = "SYSargsList"): replacement method for dependency slot for each step

updateStatus<-

signature(x = "SYSargsList"): updateStatus method for update a step status

Author

Daniela Cassol and Thomas Girke

See also

Examples

sal <- SPRproject(overwrite=TRUE)
#> Recreating directory '/home/runner/work/systemPipeR/systemPipeR/docs/reference/.SPRproject'
#> Creating file '/home/runner/work/systemPipeR/systemPipeR/docs/reference/.SPRproject/SYSargsList.yml'
targetspath <- system.file("extdata/cwl/example/targets_example.txt", 
                            package="systemPipeR")

## Constructor and `appendStep<-`
appendStep(sal) <- SYSargsList(step_name = "echo", 
                               targets=targetspath, dir=TRUE,
                               wf_file="example/workflow_example.cwl", 
                               input_file="example/example.yml", 
                               dir_path = system.file("extdata/cwl", 
                               package="systemPipeR"),
                               inputvars = c(Message = "_STRING_",
                               SampleName = "_SAMPLE_"))
                               
appendStep(sal) <- LineWise(code = {
                            hello <- lapply(getColumn(sal, step=1, 'outfiles'), function(x) yaml::read_yaml(x))
                            }, 
                            step_name = "R_read", 
                            dependency = "echo")
                   
## Accessors

stepsWF(sal)
#> $echo
#> Instance of 'SYSargs2':
#>    Slot names/accessors: 
#>       targets: 3 (M1...M3), targetsheader: 1 (lines)
#>       modules: 0
#>       wf: 1, clt: 1, yamlinput: 3 (inputs)
#>       input: 3, output: 3
#>       cmdlist: 3
#>    Sub Steps:
#>       1. echo (rendered: TRUE)
#> 
#> 
#> 
#> $R_read
#> Instance of 'LineWise'
#>     Code Chunk length: 1
#> 
statusWF(sal)
#> $echo
#> DataFrame with 3 rows and 5 columns
#>        Targets Total_Files Existing_Files Missing_Files     echo
#>    <character>   <numeric>      <numeric>     <numeric> <matrix>
#> M1          M1           1              0             1  Pending
#> M2          M2           1              0             1  Pending
#> M3          M3           1              0             1  Pending
#> 
#> $R_read
#> DataFrame with 1 row and 2 columns
#>          Step status.summary
#>   <character>    <character>
#> 1      R_read        Pending
#> 
targetsWF(sal)
#> $echo
#> DataFrame with 3 rows and 2 columns
#>                Message  SampleName
#>            <character> <character>
#> M1        Hello World!          M1
#> M2          Hello USA!          M2
#> M3 Hello Bioconductor!          M3
#> 
#> $R_read
#> DataFrame with 0 rows and 0 columns
#> 
outfiles(sal)
#> $echo
#> DataFrame with 3 rows and 1 column
#>           string
#>      <character>
#> 1 results/M1.txt
#> 2 results/M2.txt
#> 3 results/M3.txt
#> 
#> $R_read
#> DataFrame with 0 rows and 0 columns
#> 
SE(sal)
#> $echo
#> class: SummarizedExperiment 
#> dim: 0 3 
#> metadata(1): targetsheader
#> assays(0):
#> rownames: NULL
#> rowData names(0):
#> colnames(3): M1 M2 M3
#> colData names(2): Message SampleName
#> 
#> $R_read
#> NULL
#> 
dependency(sal)
#> $echo
#> [1] NA
#> 
#> $R_read
#> [1] "echo"
#> 
projectInfo(sal)
#> $project
#> [1] "/home/runner/work/systemPipeR/systemPipeR/docs/reference"
#> 
#> $data
#> [1] "data"
#> 
#> $param
#> [1] "param"
#> 
#> $results
#> [1] "results"
#> 
#> $logsDir
#> [1] ".SPRproject"
#> 
#> $sysargslist
#> [1] ".SPRproject/SYSargsList.yml"
#> 
runInfo(sal)
#> $env
#> <environment: 0x5611bc6d3a48>
#> 
#> $runOption
#> $runOption$echo
#> $runOption$echo$directory
#> [1] TRUE
#> 
#> $runOption$echo$run_step
#> [1] "mandatory"
#> 
#> $runOption$echo$run_session
#> [1] "local"
#> 
#> 
#> $runOption$R_read
#> $runOption$R_read$directory
#> [1] FALSE
#> 
#> $runOption$R_read$run_step
#> [1] "mandatory"
#> 
#> $runOption$R_read$run_session
#> [1] "local"
#> 
#> 
#> 

## Methods

baseCommand(sal, 1) ## SYSargs2 step
#> $echo
#> [1] "echo"
#> 
SampleName(sal, step="echo")
#> [1] "M1" "M2" "M3"
stepName(sal)
#> [1] "echo"   "R_read"
targetsheader(sal, step=1) ## SYSargs2 step
#> $targetsheader
#> [1] "# Project ID: SPR Example"
#> 
yamlinput(sal, step=1) ## SYSargs2 step
#> $message
#> [1] "_STRING_"
#> 
#> $SampleName
#> [1] "_SAMPLE_"
#> 
#> $results_path
#> $results_path$class
#> [1] "Directory"
#> 
#> $results_path$path
#> [1] "./results"
#> 
#> 
cmdlist(sal, step=1, targets=1:2) ## SYSargs2 step
#> $echo
#> $echo$M1
#> $echo$M1$echo
#> [1] "echo Hello World! > results/M1.txt"
#> 
#> 
#> $echo$M2
#> $echo$M2$echo
#> [1] "echo Hello USA! > results/M2.txt"
#> 
#> 
#> 
codeLine(sal, step=2) ## SYSargs2 step
#> R_read
#>     hello <- lapply(getColumn(sal, step = 1, "outfiles"), function(x) yaml::read_yaml(x))
copyEnvir(sal, list = character(), new.env = globalenv())
#> <environment: 0x5611bc6d3a48>
#> Copying to 'new.env': 
viewEnvir(sal)
#> <environment: 0x5611bc6d3a48>
#> character(0)

## Subset Methods

sal_sub <- subset(sal, subset_steps=1, input_targets=1:2, keep_steps = TRUE)
targetsIn <- getColumn(sal, step=1, position = c("outfiles"))

## Replacement
renameStep(sal, step=1) <- "new_echo"
dependency(sal, step=2) <- "new_echo"
updateColumn(sal, step=2, position = c("targetsWF")) <- data.frame(targetsIn)
targetsWF(sal)
#> $new_echo
#> DataFrame with 3 rows and 2 columns
#>                Message  SampleName
#>            <character> <character>
#> M1        Hello World!          M1
#> M2          Hello USA!          M2
#> M3 Hello Bioconductor!          M3
#> 
#> $R_read
#> DataFrame with 3 rows and 1 column
#>         targetsIn
#>       <character>
#> M1 results/M1.txt
#> M2 results/M2.txt
#> M3 results/M3.txt
#> 

replaceStep(sal, step=2) <- LineWise(code = {
                            hello <- "Printing a new message"
                            }, 
                            step_name = "R_hello", 
                            dependency = "new_echo")
#yamlinput(sal, step=1, paramName="results_path") <- list(results_path=list(
#                                            class="Directory", path="./data"))
appendCodeLine(sal, step=2, after = 0) <- "log <- log(10)"
codeLine(sal, 2)
#> R_hello
#>     log <- log(10)
#>     hello <- "Printing a new message"
replaceCodeLine(sal, step=2, line=1) <- LineWise(code = {
                            log <- log(50)
                            })
codeLine(sal, 2)
#> R_hello
#>     log <- log(50)
#>     hello <- "Printing a new message"