Class "VENNset"

Container for storing Venn intersect results created by the overLapper function. The setlist slot stores the original label sets as vectors in a list; intersectmatrix organizes the label sets in a present-absent matrix; complexitylevels represents the number of comparisons considered for each comparison set as vector of integers; and vennlist contains the Venn intersect vectors.

Objects from the Class

Objects can be created by calls of the form new("VENNset", ...).

Slots

setlist:

Object of class "list": list of vectors

intersectmatrix:

Object of class "matrix": binary matrix

complexitylevels:

Object of class "integer": vector of integers

vennlist:

Object of class "list": list of vectors

Methods

as.list

signature(x = "VENNset"): coerces VENNset to list

coerce

signature(from = "list", to = "VENNset"): as(list, "VENNset")

complexitylevels

signature(x = "VENNset"): extracts data from complexitylevels slot

intersectmatrix

signature(x = "VENNset"): extracts data from intersectmatrix slot

length

signature(x = "VENNset"): returns number of original label sets

names

signature(x = "VENNset"): extracts slot names

setlist

signature(x = "VENNset"): extracts data from setlist slot

show

signature(object = "VENNset"): summary view of VENNset objects

vennlist

signature(x = "VENNset"): extracts data from vennset slot

Author

Thomas Girke

See also

Examples

showClass("VENNset")

#> Class "VENNset" [package "systemPipeR"]
#> 
#> Slots:
#>                                                                           
#> Name:           setlist  intersectmatrix complexitylevels         vennlist
#> Class:             list           matrix          integer             list

## Sample data
setlist <- list(A=sample(letters, 18), B=sample(letters, 16),
                C=sample(letters, 20), D=sample(letters, 22),
                E=sample(letters, 18), F=sample(letters, 22))

## Create VENNset
vennset <- overLapper(setlist[1:5], type="vennsets")
class(vennset)

#> [1] "VENNset"
#> attr(,"package")
#> [1] "systemPipeR"

## Accessor methods for VENNset/INTERSECTset objects
names(vennset)

#> [1] "setlist"          "intersectmatrix"  "complexitylevels" "vennlist"        

setlist(vennset)

#> $A
#>  [1] "l" "s" "q" "j" "z" "m" "p" "g" "d" "a" "c" "y" "x" "f" "k" "o" "w" "h"
#> 
#> $B
#>  [1] "v" "q" "l" "t" "o" "r" "x" "j" "c" "d" "m" "f" "n" "b" "i" "e"
#> 
#> $C
#>  [1] "s" "r" "e" "j" "l" "d" "z" "q" "w" "i" "p" "t" "y" "o" "c" "m" "v" "a" "g"
#> [20] "h"
#> 
#> $D
#>  [1] "w" "e" "t" "k" "o" "h" "g" "s" "u" "l" "c" "r" "q" "v" "b" "x" "n" "a" "y"
#> [20] "j" "z" "i"
#> 
#> $E
#>  [1] "y" "k" "a" "p" "o" "z" "t" "x" "j" "u" "v" "d" "f" "l" "b" "r" "m" "q"
#> 

intersectmatrix(vennset)

#>   A B C D E
#> a 1 0 1 1 1
#> b 0 1 0 1 1
#> c 1 1 1 1 0
#> d 1 1 1 0 1
#> e 0 1 1 1 0
#> f 1 1 0 0 1
#> g 1 0 1 1 0
#> h 1 0 1 1 0
#> i 0 1 1 1 0
#> j 1 1 1 1 1
#> k 1 0 0 1 1
#> l 1 1 1 1 1
#> m 1 1 1 0 1
#> n 0 1 0 1 0
#> o 1 1 1 1 1
#> p 1 0 1 0 1
#> q 1 1 1 1 1
#> r 0 1 1 1 1
#> s 1 0 1 1 0
#> t 0 1 1 1 1
#> u 0 0 0 1 1
#> v 0 1 1 1 1
#> w 1 0 1 1 0
#> x 1 1 0 1 1
#> y 1 0 1 1 1
#> z 1 0 1 1 1

complexitylevels(vennset)

#>  [1] 1 1 1 1 1 2 2 2 2 2 2 2 2 2 2 3 3 3 3 3 3 3 3 3 3 4 4 4 4 4 5

vennlist(vennset)

#> $A
#> character(0)
#> 
#> $B
#> character(0)
#> 
#> $C
#> character(0)
#> 
#> $D
#> character(0)
#> 
#> $E
#> character(0)
#> 
#> $A_B
#> character(0)
#> 
#> $A_C
#> character(0)
#> 
#> $A_D
#> character(0)
#> 
#> $A_E
#> character(0)
#> 
#> $B_C
#> character(0)
#> 
#> $B_D
#> [1] "n"
#> 
#> $B_E
#> character(0)
#> 
#> $C_D
#> character(0)
#> 
#> $C_E
#> character(0)
#> 
#> $D_E
#> [1] "u"
#> 
#> $A_B_C
#> character(0)
#> 
#> $A_B_D
#> character(0)
#> 
#> $A_B_E
#> [1] "f"
#> 
#> $A_C_D
#> [1] "g" "h" "s" "w"
#> 
#> $A_C_E
#> [1] "p"
#> 
#> $A_D_E
#> [1] "k"
#> 
#> $B_C_D
#> [1] "e" "i"
#> 
#> $B_C_E
#> character(0)
#> 
#> $B_D_E
#> [1] "b"
#> 
#> $C_D_E
#> character(0)
#> 
#> $A_B_C_D
#> [1] "c"
#> 
#> $A_B_C_E
#> [1] "d" "m"
#> 
#> $A_B_D_E
#> [1] "x"
#> 
#> $A_C_D_E
#> [1] "a" "y" "z"
#> 
#> $B_C_D_E
#> [1] "r" "t" "v"
#> 
#> $A_B_C_D_E
#> [1] "j" "l" "o" "q"
#> 

## Coerce VENNset/INTERSECTset object to list
as.list(vennset)

#> $setlist
#> $setlist$A
#>  [1] "l" "s" "q" "j" "z" "m" "p" "g" "d" "a" "c" "y" "x" "f" "k" "o" "w" "h"
#> 
#> $setlist$B
#>  [1] "v" "q" "l" "t" "o" "r" "x" "j" "c" "d" "m" "f" "n" "b" "i" "e"
#> 
#> $setlist$C
#>  [1] "s" "r" "e" "j" "l" "d" "z" "q" "w" "i" "p" "t" "y" "o" "c" "m" "v" "a" "g"
#> [20] "h"
#> 
#> $setlist$D
#>  [1] "w" "e" "t" "k" "o" "h" "g" "s" "u" "l" "c" "r" "q" "v" "b" "x" "n" "a" "y"
#> [20] "j" "z" "i"
#> 
#> $setlist$E
#>  [1] "y" "k" "a" "p" "o" "z" "t" "x" "j" "u" "v" "d" "f" "l" "b" "r" "m" "q"
#> 
#> 
#> $intersectmatrix
#>   A B C D E
#> a 1 0 1 1 1
#> b 0 1 0 1 1
#> c 1 1 1 1 0
#> d 1 1 1 0 1
#> e 0 1 1 1 0
#> f 1 1 0 0 1
#> g 1 0 1 1 0
#> h 1 0 1 1 0
#> i 0 1 1 1 0
#> j 1 1 1 1 1
#> k 1 0 0 1 1
#> l 1 1 1 1 1
#> m 1 1 1 0 1
#> n 0 1 0 1 0
#> o 1 1 1 1 1
#> p 1 0 1 0 1
#> q 1 1 1 1 1
#> r 0 1 1 1 1
#> s 1 0 1 1 0
#> t 0 1 1 1 1
#> u 0 0 0 1 1
#> v 0 1 1 1 1
#> w 1 0 1 1 0
#> x 1 1 0 1 1
#> y 1 0 1 1 1
#> z 1 0 1 1 1
#> 
#> $complexitylevels
#>  [1] 1 1 1 1 1 2 2 2 2 2 2 2 2 2 2 3 3 3 3 3 3 3 3 3 3 4 4 4 4 4 5
#> 
#> $vennlist
#> $vennlist$A
#> character(0)
#> 
#> $vennlist$B
#> character(0)
#> 
#> $vennlist$C
#> character(0)
#> 
#> $vennlist$D
#> character(0)
#> 
#> $vennlist$E
#> character(0)
#> 
#> $vennlist$A_B
#> character(0)
#> 
#> $vennlist$A_C
#> character(0)
#> 
#> $vennlist$A_D
#> character(0)
#> 
#> $vennlist$A_E
#> character(0)
#> 
#> $vennlist$B_C
#> character(0)
#> 
#> $vennlist$B_D
#> [1] "n"
#> 
#> $vennlist$B_E
#> character(0)
#> 
#> $vennlist$C_D
#> character(0)
#> 
#> $vennlist$C_E
#> character(0)
#> 
#> $vennlist$D_E
#> [1] "u"
#> 
#> $vennlist$A_B_C
#> character(0)
#> 
#> $vennlist$A_B_D
#> character(0)
#> 
#> $vennlist$A_B_E
#> [1] "f"
#> 
#> $vennlist$A_C_D
#> [1] "g" "h" "s" "w"
#> 
#> $vennlist$A_C_E
#> [1] "p"
#> 
#> $vennlist$A_D_E
#> [1] "k"
#> 
#> $vennlist$B_C_D
#> [1] "e" "i"
#> 
#> $vennlist$B_C_E
#> character(0)
#> 
#> $vennlist$B_D_E
#> [1] "b"
#> 
#> $vennlist$C_D_E
#> character(0)
#> 
#> $vennlist$A_B_C_D
#> [1] "c"
#> 
#> $vennlist$A_B_C_E
#> [1] "d" "m"
#> 
#> $vennlist$A_B_D_E
#> [1] "x"
#> 
#> $vennlist$A_C_D_E
#> [1] "a" "y" "z"
#> 
#> $vennlist$B_C_D_E
#> [1] "r" "t" "v"
#> 
#> $vennlist$A_B_C_D_E
#> [1] "j" "l" "o" "q"
#> 
#>