List/check the existence of command-line tools of a workflow

These functions list/check whether required command-line tools/modules are installed in the PATH and are callable.

listCmdTools(sal, check_path = FALSE, check_module = FALSE)

listCmdModules(sal, check_module = FALSE)

Arguments

sal

SPR workflow object in SYSargsList class.

check_path

logical, whether to check if the required tools are in PATH.

check_module

logical, whether to check if the required modules are installed.

Details

Both functions by default will not check the existence of tools or modules. The default is to list the requirement.

Value

Both functions print out the list/check results as dataframe. The first column is workflow step names that require certain tools/modules. The second column is the tool/module names. The third column is logical, TRUE for the existence of the tool in PATH/modular system, if check_path = TRUE or check_module = TRUE. Otherwise, the third column will be NA.

In the case of both check_path = TRUE, check_module = TRUE for listCmdTools, the returned dataframe is still results for tool PATH checking but not module checking results. If one wish to obtain the module checking results, please use listCmdModules.

When the current workflow has no command-line (SYSargs) step, or there is no module required, or there is no modular system installed, the return will be NULL.

These two functions are automatically performed when importWF is called.

See also

Author

Le Zhang

Examples

# See examples of `importWF`