Submit command-line tools to cluster

Note: This function as been deprecated. Please use clusterRun instead. qsubRun submits command-line tools to queue (e.g. Torque) of compute cluster using run specifications defined by runX and getQsubargs functions.

qsubRun(appfct="runCommandline(args=args, runid='01')", args, qsubargs, Nqsubs = 1, package = "systemPipeR", shebang="#!/bin/bash")

Arguments

appfct
Accpets runX functions, such as appfct="runCommandline(args, runid)"

args
Argument list returned by systemArgs().

qsubargs
Argument list returned by getQsubargs().

Nqsubs
Interger defining the number of qsub processes. Note: the function will not assign more qsub processes than there are FASTQ files. E.g. if there are 10 FASTQ files and Nqsubs=20 then the function will generate only 10 qsub processes. To increase the number of CPU cores used by each process, one can increase the p value under systemArgs().

package
Package to load. Name provided as character vector of length one. Default is sytemPipeR.

shebang
defines shebang (fist line) used in submission shell script; default is set to #!/bin/bash.

Value

Returns list where list components contain FASTQ file names and their names are the qsub process IDs assiged by the queuing system. In addition, three files will be generated for each qsub submission process: submitargs0X (R object containing appargs), submitargs0X.R (R script using appargs) and submitargs0X.sh (shell submission script). In addition, the chosen runX function will output a submitargs0X_log file for each qsub process containing the executable commands processed by each qsub instance.

Author

Thomas Girke

Examples

## Construct SYSargs object from param and targets files 
param <- system.file("extdata", "tophat.param", package="systemPipeR")
targets <- system.file("extdata", "targets.txt", package="systemPipeR")
args <- systemArgs(sysma=param, mytargets=targets)

#> Warning: path[1]="./data/SRR446027_1.fastq.gz": No such file or directory

#> Warning: path[2]="./data/SRR446028_1.fastq.gz": No such file or directory

#> Warning: path[3]="./data/SRR446029_1.fastq.gz": No such file or directory

#> Warning: path[4]="./data/SRR446030_1.fastq.gz": No such file or directory

#> Warning: path[5]="./data/SRR446031_1.fastq.gz": No such file or directory

#> Warning: path[6]="./data/SRR446032_1.fastq.gz": No such file or directory

#> Warning: path[7]="./data/SRR446033_1.fastq.gz": No such file or directory

#> Warning: path[8]="./data/SRR446034_1.fastq.gz": No such file or directory

#> Warning: path[9]="./data/SRR446035_1.fastq.gz": No such file or directory

#> Warning: path[10]="./data/SRR446036_1.fastq.gz": No such file or directory

#> Warning: path[11]="./data/SRR446037_1.fastq.gz": No such file or directory

#> Warning: path[12]="./data/SRR446038_1.fastq.gz": No such file or directory

#> Warning: path[13]="./data/SRR446039_1.fastq.gz": No such file or directory

#> Warning: path[14]="./data/SRR446040_1.fastq.gz": No such file or directory

#> Warning: path[15]="./data/SRR446041_1.fastq.gz": No such file or directory

#> Warning: path[16]="./data/SRR446042_1.fastq.gz": No such file or directory

#> Warning: path[17]="./data/SRR446043_1.fastq.gz": No such file or directory

#> Warning: path[18]="./data/SRR446044_1.fastq.gz": No such file or directory

args

#> An instance of 'SYSargs' for running 'tophat' on 18 samples 

names(args); modules(args); cores(args); outpaths(args); sysargs(args)

#>  [1] "targetsin"     "targetsout"    "targetsheader" "modules"      
#>  [5] "software"      "cores"         "other"         "reference"    
#>  [9] "results"       "infile1"       "infile2"       "outfile1"     
#> [13] "sysargs"       "outpaths"     

#> [1] "bowtie2/2.2.5" "tophat/2.0.14"

#> [1] 4

#>                                                                                                              M1A 
#> "/home/runner/work/systemPipeR/systemPipeR/docs/reference/results/SRR446027_1.fastq.gz.tophat/accepted_hits.bam" 
#>                                                                                                              M1B 
#> "/home/runner/work/systemPipeR/systemPipeR/docs/reference/results/SRR446028_1.fastq.gz.tophat/accepted_hits.bam" 
#>                                                                                                              A1A 
#> "/home/runner/work/systemPipeR/systemPipeR/docs/reference/results/SRR446029_1.fastq.gz.tophat/accepted_hits.bam" 
#>                                                                                                              A1B 
#> "/home/runner/work/systemPipeR/systemPipeR/docs/reference/results/SRR446030_1.fastq.gz.tophat/accepted_hits.bam" 
#>                                                                                                              V1A 
#> "/home/runner/work/systemPipeR/systemPipeR/docs/reference/results/SRR446031_1.fastq.gz.tophat/accepted_hits.bam" 
#>                                                                                                              V1B 
#> "/home/runner/work/systemPipeR/systemPipeR/docs/reference/results/SRR446032_1.fastq.gz.tophat/accepted_hits.bam" 
#>                                                                                                              M6A 
#> "/home/runner/work/systemPipeR/systemPipeR/docs/reference/results/SRR446033_1.fastq.gz.tophat/accepted_hits.bam" 
#>                                                                                                              M6B 
#> "/home/runner/work/systemPipeR/systemPipeR/docs/reference/results/SRR446034_1.fastq.gz.tophat/accepted_hits.bam" 
#>                                                                                                              A6A 
#> "/home/runner/work/systemPipeR/systemPipeR/docs/reference/results/SRR446035_1.fastq.gz.tophat/accepted_hits.bam" 
#>                                                                                                              A6B 
#> "/home/runner/work/systemPipeR/systemPipeR/docs/reference/results/SRR446036_1.fastq.gz.tophat/accepted_hits.bam" 
#>                                                                                                              V6A 
#> "/home/runner/work/systemPipeR/systemPipeR/docs/reference/results/SRR446037_1.fastq.gz.tophat/accepted_hits.bam" 
#>                                                                                                              V6B 
#> "/home/runner/work/systemPipeR/systemPipeR/docs/reference/results/SRR446038_1.fastq.gz.tophat/accepted_hits.bam" 
#>                                                                                                             M12A 
#> "/home/runner/work/systemPipeR/systemPipeR/docs/reference/results/SRR446039_1.fastq.gz.tophat/accepted_hits.bam" 
#>                                                                                                             M12B 
#> "/home/runner/work/systemPipeR/systemPipeR/docs/reference/results/SRR446040_1.fastq.gz.tophat/accepted_hits.bam" 
#>                                                                                                             A12A 
#> "/home/runner/work/systemPipeR/systemPipeR/docs/reference/results/SRR446041_1.fastq.gz.tophat/accepted_hits.bam" 
#>                                                                                                             A12B 
#> "/home/runner/work/systemPipeR/systemPipeR/docs/reference/results/SRR446042_1.fastq.gz.tophat/accepted_hits.bam" 
#>                                                                                                             V12A 
#> "/home/runner/work/systemPipeR/systemPipeR/docs/reference/results/SRR446043_1.fastq.gz.tophat/accepted_hits.bam" 
#>                                                                                                             V12B 
#> "/home/runner/work/systemPipeR/systemPipeR/docs/reference/results/SRR446044_1.fastq.gz.tophat/accepted_hits.bam" 

#>                                                                                                                                                                                                                                                          M1A 
#> "tophat -p 4 -g 1 --segment-length 25 -i 30 -I 3000 -o /home/runner/work/systemPipeR/systemPipeR/docs/reference/results/SRR446027_1.fastq.gz.tophat /home/runner/work/systemPipeR/systemPipeR/docs/reference/data/tair10.fasta ./data/SRR446027_1.fastq.gz " 
#>                                                                                                                                                                                                                                                          M1B 
#> "tophat -p 4 -g 1 --segment-length 25 -i 30 -I 3000 -o /home/runner/work/systemPipeR/systemPipeR/docs/reference/results/SRR446028_1.fastq.gz.tophat /home/runner/work/systemPipeR/systemPipeR/docs/reference/data/tair10.fasta ./data/SRR446028_1.fastq.gz " 
#>                                                                                                                                                                                                                                                          A1A 
#> "tophat -p 4 -g 1 --segment-length 25 -i 30 -I 3000 -o /home/runner/work/systemPipeR/systemPipeR/docs/reference/results/SRR446029_1.fastq.gz.tophat /home/runner/work/systemPipeR/systemPipeR/docs/reference/data/tair10.fasta ./data/SRR446029_1.fastq.gz " 
#>                                                                                                                                                                                                                                                          A1B 
#> "tophat -p 4 -g 1 --segment-length 25 -i 30 -I 3000 -o /home/runner/work/systemPipeR/systemPipeR/docs/reference/results/SRR446030_1.fastq.gz.tophat /home/runner/work/systemPipeR/systemPipeR/docs/reference/data/tair10.fasta ./data/SRR446030_1.fastq.gz " 
#>                                                                                                                                                                                                                                                          V1A 
#> "tophat -p 4 -g 1 --segment-length 25 -i 30 -I 3000 -o /home/runner/work/systemPipeR/systemPipeR/docs/reference/results/SRR446031_1.fastq.gz.tophat /home/runner/work/systemPipeR/systemPipeR/docs/reference/data/tair10.fasta ./data/SRR446031_1.fastq.gz " 
#>                                                                                                                                                                                                                                                          V1B 
#> "tophat -p 4 -g 1 --segment-length 25 -i 30 -I 3000 -o /home/runner/work/systemPipeR/systemPipeR/docs/reference/results/SRR446032_1.fastq.gz.tophat /home/runner/work/systemPipeR/systemPipeR/docs/reference/data/tair10.fasta ./data/SRR446032_1.fastq.gz " 
#>                                                                                                                                                                                                                                                          M6A 
#> "tophat -p 4 -g 1 --segment-length 25 -i 30 -I 3000 -o /home/runner/work/systemPipeR/systemPipeR/docs/reference/results/SRR446033_1.fastq.gz.tophat /home/runner/work/systemPipeR/systemPipeR/docs/reference/data/tair10.fasta ./data/SRR446033_1.fastq.gz " 
#>                                                                                                                                                                                                                                                          M6B 
#> "tophat -p 4 -g 1 --segment-length 25 -i 30 -I 3000 -o /home/runner/work/systemPipeR/systemPipeR/docs/reference/results/SRR446034_1.fastq.gz.tophat /home/runner/work/systemPipeR/systemPipeR/docs/reference/data/tair10.fasta ./data/SRR446034_1.fastq.gz " 
#>                                                                                                                                                                                                                                                          A6A 
#> "tophat -p 4 -g 1 --segment-length 25 -i 30 -I 3000 -o /home/runner/work/systemPipeR/systemPipeR/docs/reference/results/SRR446035_1.fastq.gz.tophat /home/runner/work/systemPipeR/systemPipeR/docs/reference/data/tair10.fasta ./data/SRR446035_1.fastq.gz " 
#>                                                                                                                                                                                                                                                          A6B 
#> "tophat -p 4 -g 1 --segment-length 25 -i 30 -I 3000 -o /home/runner/work/systemPipeR/systemPipeR/docs/reference/results/SRR446036_1.fastq.gz.tophat /home/runner/work/systemPipeR/systemPipeR/docs/reference/data/tair10.fasta ./data/SRR446036_1.fastq.gz " 
#>                                                                                                                                                                                                                                                          V6A 
#> "tophat -p 4 -g 1 --segment-length 25 -i 30 -I 3000 -o /home/runner/work/systemPipeR/systemPipeR/docs/reference/results/SRR446037_1.fastq.gz.tophat /home/runner/work/systemPipeR/systemPipeR/docs/reference/data/tair10.fasta ./data/SRR446037_1.fastq.gz " 
#>                                                                                                                                                                                                                                                          V6B 
#> "tophat -p 4 -g 1 --segment-length 25 -i 30 -I 3000 -o /home/runner/work/systemPipeR/systemPipeR/docs/reference/results/SRR446038_1.fastq.gz.tophat /home/runner/work/systemPipeR/systemPipeR/docs/reference/data/tair10.fasta ./data/SRR446038_1.fastq.gz " 
#>                                                                                                                                                                                                                                                         M12A 
#> "tophat -p 4 -g 1 --segment-length 25 -i 30 -I 3000 -o /home/runner/work/systemPipeR/systemPipeR/docs/reference/results/SRR446039_1.fastq.gz.tophat /home/runner/work/systemPipeR/systemPipeR/docs/reference/data/tair10.fasta ./data/SRR446039_1.fastq.gz " 
#>                                                                                                                                                                                                                                                         M12B 
#> "tophat -p 4 -g 1 --segment-length 25 -i 30 -I 3000 -o /home/runner/work/systemPipeR/systemPipeR/docs/reference/results/SRR446040_1.fastq.gz.tophat /home/runner/work/systemPipeR/systemPipeR/docs/reference/data/tair10.fasta ./data/SRR446040_1.fastq.gz " 
#>                                                                                                                                                                                                                                                         A12A 
#> "tophat -p 4 -g 1 --segment-length 25 -i 30 -I 3000 -o /home/runner/work/systemPipeR/systemPipeR/docs/reference/results/SRR446041_1.fastq.gz.tophat /home/runner/work/systemPipeR/systemPipeR/docs/reference/data/tair10.fasta ./data/SRR446041_1.fastq.gz " 
#>                                                                                                                                                                                                                                                         A12B 
#> "tophat -p 4 -g 1 --segment-length 25 -i 30 -I 3000 -o /home/runner/work/systemPipeR/systemPipeR/docs/reference/results/SRR446042_1.fastq.gz.tophat /home/runner/work/systemPipeR/systemPipeR/docs/reference/data/tair10.fasta ./data/SRR446042_1.fastq.gz " 
#>                                                                                                                                                                                                                                                         V12A 
#> "tophat -p 4 -g 1 --segment-length 25 -i 30 -I 3000 -o /home/runner/work/systemPipeR/systemPipeR/docs/reference/results/SRR446043_1.fastq.gz.tophat /home/runner/work/systemPipeR/systemPipeR/docs/reference/data/tair10.fasta ./data/SRR446043_1.fastq.gz " 
#>                                                                                                                                                                                                                                                         V12B 
#> "tophat -p 4 -g 1 --segment-length 25 -i 30 -I 3000 -o /home/runner/work/systemPipeR/systemPipeR/docs/reference/results/SRR446044_1.fastq.gz.tophat /home/runner/work/systemPipeR/systemPipeR/docs/reference/data/tair10.fasta ./data/SRR446044_1.fastq.gz " 

if (FALSE) {
## Execute SYSargs on single machine
runCommandline(args=args)

## Execute SYSargs on multiple machines  
qsubargs <- getQsubargs(queue="batch", Nnodes="nodes=1", cores=cores(tophat), memory="mem=10gb", time="walltime=20:00:00")
qsubRun(args=args, qsubargs=qsubargs, Nqsubs=1, package="systemPipeR")
## Alignment stats
read_statsDF <- alignStats(fqpaths=tophatargs$infile1, bampaths=bampaths, fqgz=TRUE) 
read_statsDF <- cbind(read_statsDF[targets$FileName,], targets)
write.table(read_statsDF, "results/alignStats.xls", row.names=FALSE, quote=FALSE, sep="\t")
}