Render RMarkdown Report

Render the technical report file to the specified output format using pandoc.

renderReport(sysargs, fileName ="spr_report",
              rmd_title = "SPR workflow Template - Report",
              rmd_author = "Author",
              rmd_date= "Last update: `r format(Sys.time(), '%d %B, %Y')`",
              type = c("html_document"),
              desc = "This is a workflow template.",
              quiet = FALSE, open_file = TRUE)

Arguments

sysargs
object of class SYSargsList.

fileName
character string naming a file output. Default is "spr_report.Rmd".

rmd_title
string, title of the Rmd.

rmd_author
string, author(s) of the Rmd, put all authors in a single character string.

rmd_date
string, date header of Rmd.

type
The R Markdown output format to convert to. The option can be the name of a format (e.g. "pdf_document" or "html_document").

desc
string, or character vector of strings, some description text in format Rmarkdown that will be added to the document before the workflow steps start. It can be a single line or multiple lines by providing a character vector, each item is one line.

quiet
If set to TRUE, all messages returned by the function will be suppressed.

open_file
Default is TRUE.

Value

It will return an SYSargsList updated, with the file path location.

Author

Daniela Cassol

See also

Examples

sal <- SPRproject(overwrite = TRUE)
#> Recreating directory '/home/runner/work/systemPipeR/systemPipeR/docs/reference/.SPRproject'
#> Creating file '/home/runner/work/systemPipeR/systemPipeR/docs/reference/.SPRproject/SYSargsList.yml'
file_path <- system.file("extdata", "spr_simple_wf.Rmd", package = "systemPipeR")
sal <- importWF(sal, file_path = file_path, verbose = FALSE)
targetspath <- system.file("extdata/cwl/example/targets_example.txt", package = "systemPipeR")
appendStep(sal) <- SYSargsList(step_name = "echo", 
                              targets = targetspath, dir = TRUE,
                              wf_file = "example/workflow_example.cwl", input_file = "example/example.yml", 
                              dir_path = system.file("extdata/cwl", package = "systemPipeR"),
                              inputvars = c(Message = "_STRING_", SampleName = "_SAMPLE_"))
sal <- renderReport(sal, open_file = FALSE)
#> 
#> Attaching package: ‘BiocStyle’
#> The following object is masked from ‘package:systemPipeR’:
#> 
#>     output
#> 
#> 
#> processing file: spr_report.Rmd
#>   |                                                                              |                                                                      |   0%  |                                                                              |.....                                                                 |   7%
#>    inline R code fragments
#> 
#>   |                                                                              |.........                                                             |  13%
#> label: setup (with options) 
#> List of 1
#>  $ include: logi FALSE
#> 
#>   |                                                                              |..............                                                        |  20%
#>   ordinary text without R code
#> 
#>   |                                                                              |...................                                                   |  27%
#> label: load_library (with options) 
#> List of 4
#>  $ eval   : logi FALSE
#>  $ spr    : chr "r"
#>  $ spr.req: chr "mandatory"
#>  $ spr.ses: chr "local"
#> 
#>   |                                                                              |.......................                                               |  33%
#>   ordinary text without R code
#> 
#>   |                                                                              |............................                                          |  40%
#> label: export_iris (with options) 
#> List of 4
#>  $ eval   : logi FALSE
#>  $ spr    : chr "r"
#>  $ spr.req: chr "mandatory"
#>  $ spr.ses: chr "local"
#> 
#>   |                                                                              |.................................                                     |  47%
#>   ordinary text without R code
#> 
#>   |                                                                              |.....................................                                 |  53%
#> label: gzip (with options) 
#> List of 5
#>  $ eval   : logi FALSE
#>  $ spr    : chr "sysargs"
#>  $ spr.dep: chr "export_iris"
#>  $ spr.req: chr "mandatory"
#>  $ spr.ses: chr "local"
#> 
#>   |                                                                              |..........................................                            |  60%
#>   ordinary text without R code
#> 
#>   |                                                                              |...............................................                       |  67%
#> label: gunzip (with options) 
#> List of 5
#>  $ eval   : logi FALSE
#>  $ spr    : chr "sysargs"
#>  $ spr.dep: chr "gzip"
#>  $ spr.req: chr "mandatory"
#>  $ spr.ses: chr "local"
#> 
#>   |                                                                              |...................................................                   |  73%
#>   ordinary text without R code
#> 
#>   |                                                                              |........................................................              |  80%
#> label: stats (with options) 
#> List of 5
#>  $ eval   : logi FALSE
#>  $ spr    : chr "r"
#>  $ spr.dep: chr "gunzip"
#>  $ spr.req: chr "optional"
#>  $ spr.ses: chr "local"
#> 
#>   |                                                                              |.............................................................         |  87%
#>   ordinary text without R code
#> 
#>   |                                                                              |.................................................................     |  93%
#> label: echo (with options) 
#> List of 4
#>  $ eval   : logi FALSE
#>  $ spr    : chr "sysargs"
#>  $ spr.req: chr "mandatory"
#>  $ spr.ses: chr "local"
#> 
#>   |                                                                              |......................................................................| 100%
#>   ordinary text without R code
#> 
#> 
#> output file: spr_report.knit.md
#> /usr/bin/pandoc +RTS -K512m -RTS spr_report.knit.md --to html4 --from markdown+autolink_bare_uris+tex_math_single_backslash --output spr_report.html --lua-filter /home/runner/work/_temp/Library/bookdown/rmarkdown/lua/custom-environment.lua --lua-filter /home/runner/work/_temp/Library/rmarkdown/rmarkdown/lua/pagebreak.lua --lua-filter /home/runner/work/_temp/Library/rmarkdown/rmarkdown/lua/latex-div.lua --self-contained --wrap preserve --variable bs3=TRUE --standalone --section-divs --table-of-contents --toc-depth 3 --variable toc_float=1 --variable toc_selectors=h1,h2,h3 --variable toc_collapsed=1 --variable toc_smooth_scroll=1 --variable toc_print=1 --template /tmp/RtmpN8lIiZ/BiocStyle/template.html --no-highlight --variable highlightjs=1 --number-sections --variable theme=bootstrap --css /home/runner/work/_temp/Library/BiocStyle/resources/html/bioconductor.css --include-in-header /tmp/RtmpN8lIiZ/rmarkdown-str32901a6d3924.html --mathjax --variable 'mathjax-url:https://mathjax.rstudio.com/latest/MathJax.js?config=TeX-AMS-MML_HTMLorMML' --variable code_folding=show --variable code_menu=1 
#> 
#> Output created: spr_report.html
#> Success! Report created at spr_report.html