Execute SYSargs and SYSargs2

Function to execute system parameters specified in SYSargs and SYSargs2 object.

runCommandline(args, runid = "01", make_bam = TRUE, del_sam=TRUE, dir = FALSE, dir.name = NULL, force=FALSE, ...)

Arguments

args
object of class SYSargs or SYSargs2.

runid
Run identifier used for log file to track system call commands. Default is "01".

make_bam
Auto detects SAM file outputs and converts them to sorted and indexed BAM files. Default is make_bam=TRUE.

del_sam
This option allows deleting the SAM files created when the make_BAM converts the SAM files to sorted and indexed BAM files. Default is del_sam=TRUE.

dir
This option allows creating an exclusive results folder for each step in the workflow and a sub-folder for each sample defined in the targets file. All the outputs and log files for the particular step will be created in the respective folders. Default is dir=FALSE. Option available only for an object of class SYSargs2.

dir.name
Name of the workflow directory. Default is dir.name=FALSE. Note: This argument is required when the dir=TRUE.

force
Internally, the function checks if the expected output files exist, and it skips the command lines when the respective files exist. If the argument force is set to TRUE, the command line will be executed and the files overwrite. Default is force=FALSE.

...
Additional arguments to pass on to runCommandline().

Value

Output files, their paths can be obtained with outpaths() from SYSargs container or output() from SYSargs2. In addition, a character vector is returned containing the same paths.

Author

Daniela Cassol and Thomas Girke

Examples

#########################################
## Examples with \code{SYSargs} object ##
#########################################
## Construct SYSargs object from param and targets files 
param <- system.file("extdata", "hisat2.param", package="systemPipeR")
targets <- system.file("extdata", "targets.txt", package="systemPipeR")
args <- systemArgs(sysma=param, mytargets=targets)

#> Warning: path[1]="./data/SRR446027_1.fastq.gz": No such file or directory

#> Warning: path[2]="./data/SRR446028_1.fastq.gz": No such file or directory

#> Warning: path[3]="./data/SRR446029_1.fastq.gz": No such file or directory

#> Warning: path[4]="./data/SRR446030_1.fastq.gz": No such file or directory

#> Warning: path[5]="./data/SRR446031_1.fastq.gz": No such file or directory

#> Warning: path[6]="./data/SRR446032_1.fastq.gz": No such file or directory

#> Warning: path[7]="./data/SRR446033_1.fastq.gz": No such file or directory

#> Warning: path[8]="./data/SRR446034_1.fastq.gz": No such file or directory

#> Warning: path[9]="./data/SRR446035_1.fastq.gz": No such file or directory

#> Warning: path[10]="./data/SRR446036_1.fastq.gz": No such file or directory

#> Warning: path[11]="./data/SRR446037_1.fastq.gz": No such file or directory

#> Warning: path[12]="./data/SRR446038_1.fastq.gz": No such file or directory

#> Warning: path[13]="./data/SRR446039_1.fastq.gz": No such file or directory

#> Warning: path[14]="./data/SRR446040_1.fastq.gz": No such file or directory

#> Warning: path[15]="./data/SRR446041_1.fastq.gz": No such file or directory

#> Warning: path[16]="./data/SRR446042_1.fastq.gz": No such file or directory

#> Warning: path[17]="./data/SRR446043_1.fastq.gz": No such file or directory

#> Warning: path[18]="./data/SRR446044_1.fastq.gz": No such file or directory

args

#> An instance of 'SYSargs' for running 'hisat2' on 18 samples 

names(args); modules(args); cores(args); outpaths(args); sysargs(args)

#>  [1] "targetsin"     "targetsout"    "targetsheader" "modules"      
#>  [5] "software"      "cores"         "other"         "reference"    
#>  [9] "results"       "infile1"       "infile2"       "outfile1"     
#> [13] "sysargs"       "outpaths"     

#> [1] "hisat2/2.1.0"

#> [1] 4

#>                                                                                                                    M1A 
#> "/home/dcassol/danielac@ucr.edu/projects/SPR_WF_org/systemPipeR/docs/reference/results/SRR446027_1.fastq.gz.hisat.bam" 
#>                                                                                                                    M1B 
#> "/home/dcassol/danielac@ucr.edu/projects/SPR_WF_org/systemPipeR/docs/reference/results/SRR446028_1.fastq.gz.hisat.bam" 
#>                                                                                                                    A1A 
#> "/home/dcassol/danielac@ucr.edu/projects/SPR_WF_org/systemPipeR/docs/reference/results/SRR446029_1.fastq.gz.hisat.bam" 
#>                                                                                                                    A1B 
#> "/home/dcassol/danielac@ucr.edu/projects/SPR_WF_org/systemPipeR/docs/reference/results/SRR446030_1.fastq.gz.hisat.bam" 
#>                                                                                                                    V1A 
#> "/home/dcassol/danielac@ucr.edu/projects/SPR_WF_org/systemPipeR/docs/reference/results/SRR446031_1.fastq.gz.hisat.bam" 
#>                                                                                                                    V1B 
#> "/home/dcassol/danielac@ucr.edu/projects/SPR_WF_org/systemPipeR/docs/reference/results/SRR446032_1.fastq.gz.hisat.bam" 
#>                                                                                                                    M6A 
#> "/home/dcassol/danielac@ucr.edu/projects/SPR_WF_org/systemPipeR/docs/reference/results/SRR446033_1.fastq.gz.hisat.bam" 
#>                                                                                                                    M6B 
#> "/home/dcassol/danielac@ucr.edu/projects/SPR_WF_org/systemPipeR/docs/reference/results/SRR446034_1.fastq.gz.hisat.bam" 
#>                                                                                                                    A6A 
#> "/home/dcassol/danielac@ucr.edu/projects/SPR_WF_org/systemPipeR/docs/reference/results/SRR446035_1.fastq.gz.hisat.bam" 
#>                                                                                                                    A6B 
#> "/home/dcassol/danielac@ucr.edu/projects/SPR_WF_org/systemPipeR/docs/reference/results/SRR446036_1.fastq.gz.hisat.bam" 
#>                                                                                                                    V6A 
#> "/home/dcassol/danielac@ucr.edu/projects/SPR_WF_org/systemPipeR/docs/reference/results/SRR446037_1.fastq.gz.hisat.bam" 
#>                                                                                                                    V6B 
#> "/home/dcassol/danielac@ucr.edu/projects/SPR_WF_org/systemPipeR/docs/reference/results/SRR446038_1.fastq.gz.hisat.bam" 
#>                                                                                                                   M12A 
#> "/home/dcassol/danielac@ucr.edu/projects/SPR_WF_org/systemPipeR/docs/reference/results/SRR446039_1.fastq.gz.hisat.bam" 
#>                                                                                                                   M12B 
#> "/home/dcassol/danielac@ucr.edu/projects/SPR_WF_org/systemPipeR/docs/reference/results/SRR446040_1.fastq.gz.hisat.bam" 
#>                                                                                                                   A12A 
#> "/home/dcassol/danielac@ucr.edu/projects/SPR_WF_org/systemPipeR/docs/reference/results/SRR446041_1.fastq.gz.hisat.bam" 
#>                                                                                                                   A12B 
#> "/home/dcassol/danielac@ucr.edu/projects/SPR_WF_org/systemPipeR/docs/reference/results/SRR446042_1.fastq.gz.hisat.bam" 
#>                                                                                                                   V12A 
#> "/home/dcassol/danielac@ucr.edu/projects/SPR_WF_org/systemPipeR/docs/reference/results/SRR446043_1.fastq.gz.hisat.bam" 
#>                                                                                                                   V12B 
#> "/home/dcassol/danielac@ucr.edu/projects/SPR_WF_org/systemPipeR/docs/reference/results/SRR446044_1.fastq.gz.hisat.bam" 

#>                                                                                                                                                                                                                                                                                                                M1A 
#> "hisat2 -p 4 -k 1 --min-intronlen 30 --max-intronlen 3000 -S /home/dcassol/danielac@ucr.edu/projects/SPR_WF_org/systemPipeR/docs/reference/results/SRR446027_1.fastq.gz.hisat.sam /home/dcassol/danielac@ucr.edu/projects/SPR_WF_org/systemPipeR/docs/reference/data/tair10.fasta -U ./data/SRR446027_1.fastq.gz " 
#>                                                                                                                                                                                                                                                                                                                M1B 
#> "hisat2 -p 4 -k 1 --min-intronlen 30 --max-intronlen 3000 -S /home/dcassol/danielac@ucr.edu/projects/SPR_WF_org/systemPipeR/docs/reference/results/SRR446028_1.fastq.gz.hisat.sam /home/dcassol/danielac@ucr.edu/projects/SPR_WF_org/systemPipeR/docs/reference/data/tair10.fasta -U ./data/SRR446028_1.fastq.gz " 
#>                                                                                                                                                                                                                                                                                                                A1A 
#> "hisat2 -p 4 -k 1 --min-intronlen 30 --max-intronlen 3000 -S /home/dcassol/danielac@ucr.edu/projects/SPR_WF_org/systemPipeR/docs/reference/results/SRR446029_1.fastq.gz.hisat.sam /home/dcassol/danielac@ucr.edu/projects/SPR_WF_org/systemPipeR/docs/reference/data/tair10.fasta -U ./data/SRR446029_1.fastq.gz " 
#>                                                                                                                                                                                                                                                                                                                A1B 
#> "hisat2 -p 4 -k 1 --min-intronlen 30 --max-intronlen 3000 -S /home/dcassol/danielac@ucr.edu/projects/SPR_WF_org/systemPipeR/docs/reference/results/SRR446030_1.fastq.gz.hisat.sam /home/dcassol/danielac@ucr.edu/projects/SPR_WF_org/systemPipeR/docs/reference/data/tair10.fasta -U ./data/SRR446030_1.fastq.gz " 
#>                                                                                                                                                                                                                                                                                                                V1A 
#> "hisat2 -p 4 -k 1 --min-intronlen 30 --max-intronlen 3000 -S /home/dcassol/danielac@ucr.edu/projects/SPR_WF_org/systemPipeR/docs/reference/results/SRR446031_1.fastq.gz.hisat.sam /home/dcassol/danielac@ucr.edu/projects/SPR_WF_org/systemPipeR/docs/reference/data/tair10.fasta -U ./data/SRR446031_1.fastq.gz " 
#>                                                                                                                                                                                                                                                                                                                V1B 
#> "hisat2 -p 4 -k 1 --min-intronlen 30 --max-intronlen 3000 -S /home/dcassol/danielac@ucr.edu/projects/SPR_WF_org/systemPipeR/docs/reference/results/SRR446032_1.fastq.gz.hisat.sam /home/dcassol/danielac@ucr.edu/projects/SPR_WF_org/systemPipeR/docs/reference/data/tair10.fasta -U ./data/SRR446032_1.fastq.gz " 
#>                                                                                                                                                                                                                                                                                                                M6A 
#> "hisat2 -p 4 -k 1 --min-intronlen 30 --max-intronlen 3000 -S /home/dcassol/danielac@ucr.edu/projects/SPR_WF_org/systemPipeR/docs/reference/results/SRR446033_1.fastq.gz.hisat.sam /home/dcassol/danielac@ucr.edu/projects/SPR_WF_org/systemPipeR/docs/reference/data/tair10.fasta -U ./data/SRR446033_1.fastq.gz " 
#>                                                                                                                                                                                                                                                                                                                M6B 
#> "hisat2 -p 4 -k 1 --min-intronlen 30 --max-intronlen 3000 -S /home/dcassol/danielac@ucr.edu/projects/SPR_WF_org/systemPipeR/docs/reference/results/SRR446034_1.fastq.gz.hisat.sam /home/dcassol/danielac@ucr.edu/projects/SPR_WF_org/systemPipeR/docs/reference/data/tair10.fasta -U ./data/SRR446034_1.fastq.gz " 
#>                                                                                                                                                                                                                                                                                                                A6A 
#> "hisat2 -p 4 -k 1 --min-intronlen 30 --max-intronlen 3000 -S /home/dcassol/danielac@ucr.edu/projects/SPR_WF_org/systemPipeR/docs/reference/results/SRR446035_1.fastq.gz.hisat.sam /home/dcassol/danielac@ucr.edu/projects/SPR_WF_org/systemPipeR/docs/reference/data/tair10.fasta -U ./data/SRR446035_1.fastq.gz " 
#>                                                                                                                                                                                                                                                                                                                A6B 
#> "hisat2 -p 4 -k 1 --min-intronlen 30 --max-intronlen 3000 -S /home/dcassol/danielac@ucr.edu/projects/SPR_WF_org/systemPipeR/docs/reference/results/SRR446036_1.fastq.gz.hisat.sam /home/dcassol/danielac@ucr.edu/projects/SPR_WF_org/systemPipeR/docs/reference/data/tair10.fasta -U ./data/SRR446036_1.fastq.gz " 
#>                                                                                                                                                                                                                                                                                                                V6A 
#> "hisat2 -p 4 -k 1 --min-intronlen 30 --max-intronlen 3000 -S /home/dcassol/danielac@ucr.edu/projects/SPR_WF_org/systemPipeR/docs/reference/results/SRR446037_1.fastq.gz.hisat.sam /home/dcassol/danielac@ucr.edu/projects/SPR_WF_org/systemPipeR/docs/reference/data/tair10.fasta -U ./data/SRR446037_1.fastq.gz " 
#>                                                                                                                                                                                                                                                                                                                V6B 
#> "hisat2 -p 4 -k 1 --min-intronlen 30 --max-intronlen 3000 -S /home/dcassol/danielac@ucr.edu/projects/SPR_WF_org/systemPipeR/docs/reference/results/SRR446038_1.fastq.gz.hisat.sam /home/dcassol/danielac@ucr.edu/projects/SPR_WF_org/systemPipeR/docs/reference/data/tair10.fasta -U ./data/SRR446038_1.fastq.gz " 
#>                                                                                                                                                                                                                                                                                                               M12A 
#> "hisat2 -p 4 -k 1 --min-intronlen 30 --max-intronlen 3000 -S /home/dcassol/danielac@ucr.edu/projects/SPR_WF_org/systemPipeR/docs/reference/results/SRR446039_1.fastq.gz.hisat.sam /home/dcassol/danielac@ucr.edu/projects/SPR_WF_org/systemPipeR/docs/reference/data/tair10.fasta -U ./data/SRR446039_1.fastq.gz " 
#>                                                                                                                                                                                                                                                                                                               M12B 
#> "hisat2 -p 4 -k 1 --min-intronlen 30 --max-intronlen 3000 -S /home/dcassol/danielac@ucr.edu/projects/SPR_WF_org/systemPipeR/docs/reference/results/SRR446040_1.fastq.gz.hisat.sam /home/dcassol/danielac@ucr.edu/projects/SPR_WF_org/systemPipeR/docs/reference/data/tair10.fasta -U ./data/SRR446040_1.fastq.gz " 
#>                                                                                                                                                                                                                                                                                                               A12A 
#> "hisat2 -p 4 -k 1 --min-intronlen 30 --max-intronlen 3000 -S /home/dcassol/danielac@ucr.edu/projects/SPR_WF_org/systemPipeR/docs/reference/results/SRR446041_1.fastq.gz.hisat.sam /home/dcassol/danielac@ucr.edu/projects/SPR_WF_org/systemPipeR/docs/reference/data/tair10.fasta -U ./data/SRR446041_1.fastq.gz " 
#>                                                                                                                                                                                                                                                                                                               A12B 
#> "hisat2 -p 4 -k 1 --min-intronlen 30 --max-intronlen 3000 -S /home/dcassol/danielac@ucr.edu/projects/SPR_WF_org/systemPipeR/docs/reference/results/SRR446042_1.fastq.gz.hisat.sam /home/dcassol/danielac@ucr.edu/projects/SPR_WF_org/systemPipeR/docs/reference/data/tair10.fasta -U ./data/SRR446042_1.fastq.gz " 
#>                                                                                                                                                                                                                                                                                                               V12A 
#> "hisat2 -p 4 -k 1 --min-intronlen 30 --max-intronlen 3000 -S /home/dcassol/danielac@ucr.edu/projects/SPR_WF_org/systemPipeR/docs/reference/results/SRR446043_1.fastq.gz.hisat.sam /home/dcassol/danielac@ucr.edu/projects/SPR_WF_org/systemPipeR/docs/reference/data/tair10.fasta -U ./data/SRR446043_1.fastq.gz " 
#>                                                                                                                                                                                                                                                                                                               V12B 
#> "hisat2 -p 4 -k 1 --min-intronlen 30 --max-intronlen 3000 -S /home/dcassol/danielac@ucr.edu/projects/SPR_WF_org/systemPipeR/docs/reference/results/SRR446044_1.fastq.gz.hisat.sam /home/dcassol/danielac@ucr.edu/projects/SPR_WF_org/systemPipeR/docs/reference/data/tair10.fasta -U ./data/SRR446044_1.fastq.gz " 

if (FALSE) {
## Execute SYSargs on single machine
runCommandline(args=args)

## Execute SYSargs on multiple machines of a compute cluster. 
file.copy(system.file("extdata", ".batchtools.conf.R", package="systemPipeR"), ".")
file.copy(system.file("extdata", "batchtools.slurm.tmpl", package="systemPipeR"), ".")
resources <- list(walltime=120, ntasks=1, ncpus=cores(args), memory=1024) 
reg <- clusterRun(args, FUN = runCommandline, conffile=".batchtools.conf.R", template="batchtools.slurm.tmpl", Njobs=18, runid="01", resourceList=resources)

## Monitor progress of submitted jobs
getStatus(reg=reg)
file.exists(outpaths(args))

## Alignment stats
read_statsDF <- alignStats(args) 
read_statsDF <- cbind(read_statsDF[targets$FileName,], targets)
write.table(read_statsDF, "results/alignStats.xls", row.names=FALSE, quote=FALSE, sep="\t")
}

##########################################
## Examples with \code{SYSargs2} object ##
##########################################
## Construct SYSargs2 object from CWl param, CWL input, and targets files 
targets <- system.file("extdata", "targets.txt", package="systemPipeR")
dir_path <- system.file("extdata/cwl/hisat2/hisat2-se", package="systemPipeR")
WF <- loadWorkflow(targets=targets, wf_file="hisat2-mapping-se.cwl", 
                  input_file="hisat2-mapping-se.yml", dir_path=dir_path)
WF <- renderWF(WF, inputvars=c(FileName="_FASTQ_PATH1_", SampleName="_SampleName_"))
WF

#> Instance of 'SYSargs2':
#>    Slot names/accessors: 
#>       targets: 18 (M1A...V12B), targetsheader: 4 (lines)
#>       modules: 1
#>       wf: 0, clt: 1, yamlinput: 7 (components)
#>       input: 18, output: 18
#>       cmdlist: 18
#>    WF Steps:
#>       1. hisat2-mapping-se (rendered: TRUE)
#> 
#> 

names(WF); modules(WF); targets(WF)[1]; cmdlist(WF)[1:2]; output(WF)

#>  [1] "targets"       "targetsheader" "modules"       "wf"           
#>  [5] "clt"           "yamlinput"     "cmdlist"       "input"        
#>  [9] "output"        "cwlfiles"      "inputvars"    

#>        module1 
#> "hisat2/2.1.0" 

#> $M1A
#> $M1A$FileName
#> [1] "./data/SRR446027_1.fastq.gz"
#> 
#> $M1A$SampleName
#> [1] "M1A"
#> 
#> $M1A$Factor
#> [1] "M1"
#> 
#> $M1A$SampleLong
#> [1] "Mock.1h.A"
#> 
#> $M1A$Experiment
#> [1] 1
#> 
#> $M1A$Date
#> [1] "23-Mar-2012"
#> 
#> 

#> $M1A
#> $M1A$`hisat2-mapping-se`
#> [1] "hisat2 -S ./results/M1A.sam  -x ./data/tair10.fasta  -k 1  --min-intronlen 30  --max-intronlen 3000  -U ./data/SRR446027_1.fastq.gz --threads 4"
#> 
#> 
#> $M1B
#> $M1B$`hisat2-mapping-se`
#> [1] "hisat2 -S ./results/M1B.sam  -x ./data/tair10.fasta  -k 1  --min-intronlen 30  --max-intronlen 3000  -U ./data/SRR446028_1.fastq.gz --threads 4"
#> 
#> 

#> $M1A
#> $M1A$`hisat2-mapping-se`
#> [1] "./results/M1A.sam"
#> 
#> 
#> $M1B
#> $M1B$`hisat2-mapping-se`
#> [1] "./results/M1B.sam"
#> 
#> 
#> $A1A
#> $A1A$`hisat2-mapping-se`
#> [1] "./results/A1A.sam"
#> 
#> 
#> $A1B
#> $A1B$`hisat2-mapping-se`
#> [1] "./results/A1B.sam"
#> 
#> 
#> $V1A
#> $V1A$`hisat2-mapping-se`
#> [1] "./results/V1A.sam"
#> 
#> 
#> $V1B
#> $V1B$`hisat2-mapping-se`
#> [1] "./results/V1B.sam"
#> 
#> 
#> $M6A
#> $M6A$`hisat2-mapping-se`
#> [1] "./results/M6A.sam"
#> 
#> 
#> $M6B
#> $M6B$`hisat2-mapping-se`
#> [1] "./results/M6B.sam"
#> 
#> 
#> $A6A
#> $A6A$`hisat2-mapping-se`
#> [1] "./results/A6A.sam"
#> 
#> 
#> $A6B
#> $A6B$`hisat2-mapping-se`
#> [1] "./results/A6B.sam"
#> 
#> 
#> $V6A
#> $V6A$`hisat2-mapping-se`
#> [1] "./results/V6A.sam"
#> 
#> 
#> $V6B
#> $V6B$`hisat2-mapping-se`
#> [1] "./results/V6B.sam"
#> 
#> 
#> $M12A
#> $M12A$`hisat2-mapping-se`
#> [1] "./results/M12A.sam"
#> 
#> 
#> $M12B
#> $M12B$`hisat2-mapping-se`
#> [1] "./results/M12B.sam"
#> 
#> 
#> $A12A
#> $A12A$`hisat2-mapping-se`
#> [1] "./results/A12A.sam"
#> 
#> 
#> $A12B
#> $A12B$`hisat2-mapping-se`
#> [1] "./results/A12B.sam"
#> 
#> 
#> $V12A
#> $V12A$`hisat2-mapping-se`
#> [1] "./results/V12A.sam"
#> 
#> 
#> $V12B
#> $V12B$`hisat2-mapping-se`
#> [1] "./results/V12B.sam"
#> 
#> 

if (FALSE) {
## Execute SYSargs2 on single machine
WF <- runCommandline(args=WF)

## Execute SYSargs on multiple machines of a compute cluster. 
file.copy(system.file("extdata", ".batchtools.conf.R", package="systemPipeR"), ".")
file.copy(system.file("extdata", "batchtools.slurm.tmpl", package="systemPipeR"), ".")
resources <- list(walltime=120, ntasks=1, ncpus=4, memory=1024) 
reg <- clusterRun(WF, FUN = runCommandline, more.args = list(args = WF, make_bam = TRUE), conffile=".batchtools.conf.R", template="batchtools.slurm.tmpl", Njobs=18, runid="01", resourceList=resources)

## Monitor progress of submitted jobs
getStatus(reg=reg)

## Updates the path in the object \code{output(WF)}
WF <- output_update(WF, dir=FALSE, replace=TRUE, extension=c(".sam", ".bam"))

## Alignment stats
read_statsDF <- alignStats(WF) 
read_statsDF <- cbind(read_statsDF[targets$FileName,], targets)
write.table(read_statsDF, "results/alignStats.xls", row.names=FALSE, quote=FALSE, sep="\t")
}