Translate SYSargsList back to a workflow template Rmarkdown file

This function takes a SYSargsList object and translate it to SPR workflow template Rmarkdown format.

sal2rmd(sal, out_path = "spr_template.Rmd", rmd_title = "SPR workflow template", 
        rmd_author = "my name", 
        rmd_date = "Last update: `r format(Sys.time(), '%d %B, %Y')`", 
        rmd_output = "html_document", 
        desc = "This is a workflow template.", verbose = TRUE)

Arguments

sal

SYSargsList object.

out_path

string, output file name.

rmd_title

string, title of the Rmd.

rmd_author

string, author(s) of the Rmd, put all authors in a single character string.

rmd_date

string, date header of Rmd.

rmd_output

string, output format of Rmd, used in header.

desc

string, or character vector of strings, some description text in format Rmarkdown that will be added to the document before the workflow steps start. It can be a single line or multiple lines by providing a character vector, each item is one line.

verbose

logical. If TRUE will show you more information as the function runs.

Value

no return

Author

Le Zhang and Daniela Cassol

Examples

file_path <- system.file("extdata/spr_simple_wf.Rmd", package="systemPipeR")
sal <- SPRproject(overwrite = TRUE)
#> Recreating directory '/home/runner/work/systemPipeR/systemPipeR/docs/reference/.SPRproject'
#> Creating file '/home/runner/work/systemPipeR/systemPipeR/docs/reference/.SPRproject/SYSargsList.yml'
sal <- importWF(sal, file_path)
#> Reading Rmd file
#> 
#>  ---- Actions ----
#> Checking chunk SPR option
#> Ignore non-SPR chunks: 17
#> Checking chunk spr.req option
#> Checking chunk spr.ses option
#> Checking chunk eval values
#> Resolve step names
#> Check duplicated step names
#> Checking chunk dependencies
#> Use the previous step as dependency for steps without 'spr.dep' options: 25, 37
#> Parse chunk code
#> ---- Succes! Create output ----
#> Now importing step 'load_library' 
#> Now importing step 'export_iris' 
#> Now importing step 'gzip' 
#> Now importing step 'gunzip' 
#> Now importing step 'stats' 
#> 
sal2rmd(sal)
#> sal2rmd starts, pre-checks...
#> Open spr_template.Rmd to write
#> Write Rmd header and description
#> Now writing step 1 load_library
#> Now writing step 2 export_iris
#> Now writing step 3 gzip
#> Now writing step 4 gunzip
#> Now writing step 5 stats
#> Success! File created at spr_template.Rmd