SYSargs2 accessor methods

Methods to access information from SYSargs2 object.

sysargs2(x)

Arguments

x
object of class SYSargs2

Value

various outputs

Author

Daniela Cassol and Thomas Girke

Examples

## Construct SYSargs2 object from CWl param, CWL input, and targets files 
targets <- system.file("extdata", "targets.txt", package="systemPipeR")
dir_path <- system.file("extdata/cwl/hisat2/hisat2-se", package="systemPipeR")
WF <- loadWorkflow(targets=targets, wf_file="hisat2-mapping-se.cwl", 
                  input_file="hisat2-mapping-se.yml", dir_path=dir_path)
WF <- renderWF(WF, inputvars=c(FileName="_FASTQ_PATH1_", SampleName="_SampleName_"))
WF

#> Instance of 'SYSargs2':
#>    Slot names/accessors: 
#>       targets: 18 (M1A...V12B), targetsheader: 4 (lines)
#>       modules: 1
#>       wf: 0, clt: 1, yamlinput: 7 (components)
#>       input: 18, output: 18
#>       cmdlist: 18
#>    WF Steps:
#>       1. hisat2-mapping-se (rendered: TRUE)
#> 
#> 

names(WF); modules(WF); targets(WF)[1]; cmdlist(WF)[1:2]; output(WF)

#>  [1] "targets"       "targetsheader" "modules"       "wf"           
#>  [5] "clt"           "yamlinput"     "cmdlist"       "input"        
#>  [9] "output"        "cwlfiles"      "inputvars"    

#>        module1 
#> "hisat2/2.1.0" 

#> $M1A
#> $M1A$FileName
#> [1] "./data/SRR446027_1.fastq.gz"
#> 
#> $M1A$SampleName
#> [1] "M1A"
#> 
#> $M1A$Factor
#> [1] "M1"
#> 
#> $M1A$SampleLong
#> [1] "Mock.1h.A"
#> 
#> $M1A$Experiment
#> [1] 1
#> 
#> $M1A$Date
#> [1] "23-Mar-2012"
#> 
#> 

#> $M1A
#> $M1A$`hisat2-mapping-se`
#> [1] "hisat2 -S ./results/M1A.sam  -x ./data/tair10.fasta  -k 1  --min-intronlen 30  --max-intronlen 3000  -U ./data/SRR446027_1.fastq.gz --threads 4"
#> 
#> 
#> $M1B
#> $M1B$`hisat2-mapping-se`
#> [1] "hisat2 -S ./results/M1B.sam  -x ./data/tair10.fasta  -k 1  --min-intronlen 30  --max-intronlen 3000  -U ./data/SRR446028_1.fastq.gz --threads 4"
#> 
#> 

#> $M1A
#> $M1A$`hisat2-mapping-se`
#> [1] "./results/M1A.sam"
#> 
#> 
#> $M1B
#> $M1B$`hisat2-mapping-se`
#> [1] "./results/M1B.sam"
#> 
#> 
#> $A1A
#> $A1A$`hisat2-mapping-se`
#> [1] "./results/A1A.sam"
#> 
#> 
#> $A1B
#> $A1B$`hisat2-mapping-se`
#> [1] "./results/A1B.sam"
#> 
#> 
#> $V1A
#> $V1A$`hisat2-mapping-se`
#> [1] "./results/V1A.sam"
#> 
#> 
#> $V1B
#> $V1B$`hisat2-mapping-se`
#> [1] "./results/V1B.sam"
#> 
#> 
#> $M6A
#> $M6A$`hisat2-mapping-se`
#> [1] "./results/M6A.sam"
#> 
#> 
#> $M6B
#> $M6B$`hisat2-mapping-se`
#> [1] "./results/M6B.sam"
#> 
#> 
#> $A6A
#> $A6A$`hisat2-mapping-se`
#> [1] "./results/A6A.sam"
#> 
#> 
#> $A6B
#> $A6B$`hisat2-mapping-se`
#> [1] "./results/A6B.sam"
#> 
#> 
#> $V6A
#> $V6A$`hisat2-mapping-se`
#> [1] "./results/V6A.sam"
#> 
#> 
#> $V6B
#> $V6B$`hisat2-mapping-se`
#> [1] "./results/V6B.sam"
#> 
#> 
#> $M12A
#> $M12A$`hisat2-mapping-se`
#> [1] "./results/M12A.sam"
#> 
#> 
#> $M12B
#> $M12B$`hisat2-mapping-se`
#> [1] "./results/M12B.sam"
#> 
#> 
#> $A12A
#> $A12A$`hisat2-mapping-se`
#> [1] "./results/A12A.sam"
#> 
#> 
#> $A12B
#> $A12B$`hisat2-mapping-se`
#> [1] "./results/A12B.sam"
#> 
#> 
#> $V12A
#> $V12A$`hisat2-mapping-se`
#> [1] "./results/V12A.sam"
#> 
#> 
#> $V12B
#> $V12B$`hisat2-mapping-se`
#> [1] "./results/V12B.sam"
#> 
#> 

sysargs2(WF)

#> $targets
#> $targets$M1A
#> $targets$M1A$FileName
#> [1] "./data/SRR446027_1.fastq.gz"
#> 
#> $targets$M1A$SampleName
#> [1] "M1A"
#> 
#> $targets$M1A$Factor
#> [1] "M1"
#> 
#> $targets$M1A$SampleLong
#> [1] "Mock.1h.A"
#> 
#> $targets$M1A$Experiment
#> [1] 1
#> 
#> $targets$M1A$Date
#> [1] "23-Mar-2012"
#> 
#> 
#> $targets$M1B
#> $targets$M1B$FileName
#> [1] "./data/SRR446028_1.fastq.gz"
#> 
#> $targets$M1B$SampleName
#> [1] "M1B"
#> 
#> $targets$M1B$Factor
#> [1] "M1"
#> 
#> $targets$M1B$SampleLong
#> [1] "Mock.1h.B"
#> 
#> $targets$M1B$Experiment
#> [1] 1
#> 
#> $targets$M1B$Date
#> [1] "23-Mar-2012"
#> 
#> 
#> $targets$A1A
#> $targets$A1A$FileName
#> [1] "./data/SRR446029_1.fastq.gz"
#> 
#> $targets$A1A$SampleName
#> [1] "A1A"
#> 
#> $targets$A1A$Factor
#> [1] "A1"
#> 
#> $targets$A1A$SampleLong
#> [1] "Avr.1h.A"
#> 
#> $targets$A1A$Experiment
#> [1] 1
#> 
#> $targets$A1A$Date
#> [1] "23-Mar-2012"
#> 
#> 
#> $targets$A1B
#> $targets$A1B$FileName
#> [1] "./data/SRR446030_1.fastq.gz"
#> 
#> $targets$A1B$SampleName
#> [1] "A1B"
#> 
#> $targets$A1B$Factor
#> [1] "A1"
#> 
#> $targets$A1B$SampleLong
#> [1] "Avr.1h.B"
#> 
#> $targets$A1B$Experiment
#> [1] 1
#> 
#> $targets$A1B$Date
#> [1] "23-Mar-2012"
#> 
#> 
#> $targets$V1A
#> $targets$V1A$FileName
#> [1] "./data/SRR446031_1.fastq.gz"
#> 
#> $targets$V1A$SampleName
#> [1] "V1A"
#> 
#> $targets$V1A$Factor
#> [1] "V1"
#> 
#> $targets$V1A$SampleLong
#> [1] "Vir.1h.A"
#> 
#> $targets$V1A$Experiment
#> [1] 1
#> 
#> $targets$V1A$Date
#> [1] "23-Mar-2012"
#> 
#> 
#> $targets$V1B
#> $targets$V1B$FileName
#> [1] "./data/SRR446032_1.fastq.gz"
#> 
#> $targets$V1B$SampleName
#> [1] "V1B"
#> 
#> $targets$V1B$Factor
#> [1] "V1"
#> 
#> $targets$V1B$SampleLong
#> [1] "Vir.1h.B"
#> 
#> $targets$V1B$Experiment
#> [1] 1
#> 
#> $targets$V1B$Date
#> [1] "23-Mar-2012"
#> 
#> 
#> $targets$M6A
#> $targets$M6A$FileName
#> [1] "./data/SRR446033_1.fastq.gz"
#> 
#> $targets$M6A$SampleName
#> [1] "M6A"
#> 
#> $targets$M6A$Factor
#> [1] "M6"
#> 
#> $targets$M6A$SampleLong
#> [1] "Mock.6h.A"
#> 
#> $targets$M6A$Experiment
#> [1] 1
#> 
#> $targets$M6A$Date
#> [1] "23-Mar-2012"
#> 
#> 
#> $targets$M6B
#> $targets$M6B$FileName
#> [1] "./data/SRR446034_1.fastq.gz"
#> 
#> $targets$M6B$SampleName
#> [1] "M6B"
#> 
#> $targets$M6B$Factor
#> [1] "M6"
#> 
#> $targets$M6B$SampleLong
#> [1] "Mock.6h.B"
#> 
#> $targets$M6B$Experiment
#> [1] 1
#> 
#> $targets$M6B$Date
#> [1] "23-Mar-2012"
#> 
#> 
#> $targets$A6A
#> $targets$A6A$FileName
#> [1] "./data/SRR446035_1.fastq.gz"
#> 
#> $targets$A6A$SampleName
#> [1] "A6A"
#> 
#> $targets$A6A$Factor
#> [1] "A6"
#> 
#> $targets$A6A$SampleLong
#> [1] "Avr.6h.A"
#> 
#> $targets$A6A$Experiment
#> [1] 1
#> 
#> $targets$A6A$Date
#> [1] "23-Mar-2012"
#> 
#> 
#> $targets$A6B
#> $targets$A6B$FileName
#> [1] "./data/SRR446036_1.fastq.gz"
#> 
#> $targets$A6B$SampleName
#> [1] "A6B"
#> 
#> $targets$A6B$Factor
#> [1] "A6"
#> 
#> $targets$A6B$SampleLong
#> [1] "Avr.6h.B"
#> 
#> $targets$A6B$Experiment
#> [1] 1
#> 
#> $targets$A6B$Date
#> [1] "23-Mar-2012"
#> 
#> 
#> $targets$V6A
#> $targets$V6A$FileName
#> [1] "./data/SRR446037_1.fastq.gz"
#> 
#> $targets$V6A$SampleName
#> [1] "V6A"
#> 
#> $targets$V6A$Factor
#> [1] "V6"
#> 
#> $targets$V6A$SampleLong
#> [1] "Vir.6h.A"
#> 
#> $targets$V6A$Experiment
#> [1] 1
#> 
#> $targets$V6A$Date
#> [1] "23-Mar-2012"
#> 
#> 
#> $targets$V6B
#> $targets$V6B$FileName
#> [1] "./data/SRR446038_1.fastq.gz"
#> 
#> $targets$V6B$SampleName
#> [1] "V6B"
#> 
#> $targets$V6B$Factor
#> [1] "V6"
#> 
#> $targets$V6B$SampleLong
#> [1] "Vir.6h.B"
#> 
#> $targets$V6B$Experiment
#> [1] 1
#> 
#> $targets$V6B$Date
#> [1] "23-Mar-2012"
#> 
#> 
#> $targets$M12A
#> $targets$M12A$FileName
#> [1] "./data/SRR446039_1.fastq.gz"
#> 
#> $targets$M12A$SampleName
#> [1] "M12A"
#> 
#> $targets$M12A$Factor
#> [1] "M12"
#> 
#> $targets$M12A$SampleLong
#> [1] "Mock.12h.A"
#> 
#> $targets$M12A$Experiment
#> [1] 1
#> 
#> $targets$M12A$Date
#> [1] "23-Mar-2012"
#> 
#> 
#> $targets$M12B
#> $targets$M12B$FileName
#> [1] "./data/SRR446040_1.fastq.gz"
#> 
#> $targets$M12B$SampleName
#> [1] "M12B"
#> 
#> $targets$M12B$Factor
#> [1] "M12"
#> 
#> $targets$M12B$SampleLong
#> [1] "Mock.12h.B"
#> 
#> $targets$M12B$Experiment
#> [1] 1
#> 
#> $targets$M12B$Date
#> [1] "23-Mar-2012"
#> 
#> 
#> $targets$A12A
#> $targets$A12A$FileName
#> [1] "./data/SRR446041_1.fastq.gz"
#> 
#> $targets$A12A$SampleName
#> [1] "A12A"
#> 
#> $targets$A12A$Factor
#> [1] "A12"
#> 
#> $targets$A12A$SampleLong
#> [1] "Avr.12h.A"
#> 
#> $targets$A12A$Experiment
#> [1] 1
#> 
#> $targets$A12A$Date
#> [1] "23-Mar-2012"
#> 
#> 
#> $targets$A12B
#> $targets$A12B$FileName
#> [1] "./data/SRR446042_1.fastq.gz"
#> 
#> $targets$A12B$SampleName
#> [1] "A12B"
#> 
#> $targets$A12B$Factor
#> [1] "A12"
#> 
#> $targets$A12B$SampleLong
#> [1] "Avr.12h.B"
#> 
#> $targets$A12B$Experiment
#> [1] 1
#> 
#> $targets$A12B$Date
#> [1] "23-Mar-2012"
#> 
#> 
#> $targets$V12A
#> $targets$V12A$FileName
#> [1] "./data/SRR446043_1.fastq.gz"
#> 
#> $targets$V12A$SampleName
#> [1] "V12A"
#> 
#> $targets$V12A$Factor
#> [1] "V12"
#> 
#> $targets$V12A$SampleLong
#> [1] "Vir.12h.A"
#> 
#> $targets$V12A$Experiment
#> [1] 1
#> 
#> $targets$V12A$Date
#> [1] "23-Mar-2012"
#> 
#> 
#> $targets$V12B
#> $targets$V12B$FileName
#> [1] "./data/SRR446044_1.fastq.gz"
#> 
#> $targets$V12B$SampleName
#> [1] "V12B"
#> 
#> $targets$V12B$Factor
#> [1] "V12"
#> 
#> $targets$V12B$SampleLong
#> [1] "Vir.12h.B"
#> 
#> $targets$V12B$Experiment
#> [1] 1
#> 
#> $targets$V12B$Date
#> [1] "23-Mar-2012"
#> 
#> 
#> 
#> $targetsheader
#> $targetsheader$targetsheader
#> [1] "# Project ID: Arabidopsis - Pseudomonas alternative splicing study (SRA: SRP010938; PMID: 24098335)"                                                                              
#> [2] "# The following line(s) allow to specify the contrasts needed for comparative analyses, such as DEG identification. All possible comparisons can be specified with 'CMPset: ALL'."
#> [3] "# <CMP> CMPset1: M1-A1, M1-V1, A1-V1, M6-A6, M6-V6, A6-V6, M12-A12, M12-V12, A12-V12"                                                                                             
#> [4] "# <CMP> CMPset2: ALL"                                                                                                                                                             
#> 
#> 
#> $modules
#> $modules$module1
#> [1] "hisat2/2.1.0"
#> 
#> 
#> $wf
#> list()
#> 
#> $clt
#> $clt$`hisat2-mapping-se.cwl`
#> $clt$`hisat2-mapping-se.cwl`$cwlVersion
#> [1] "v1.0"
#> 
#> $clt$`hisat2-mapping-se.cwl`$class
#> [1] "CommandLineTool"
#> 
#> $clt$`hisat2-mapping-se.cwl`$doc
#> [1] "[HISAT2](https://ccb.jhu.edu/software/hisat2/index.shtml): graph-based alignment of next generation sequencing reads to a population of genomes"
#> 
#> $clt$`hisat2-mapping-se.cwl`$label
#> [1] "Last updated 08/2019"
#> 
#> $clt$`hisat2-mapping-se.cwl`$hints
#> $clt$`hisat2-mapping-se.cwl`$hints$SoftwareRequirement
#> $clt$`hisat2-mapping-se.cwl`$hints$SoftwareRequirement$packages
#> $clt$`hisat2-mapping-se.cwl`$hints$SoftwareRequirement$packages[[1]]
#> $clt$`hisat2-mapping-se.cwl`$hints$SoftwareRequirement$packages[[1]]$package
#> [1] "hisat2"
#> 
#> $clt$`hisat2-mapping-se.cwl`$hints$SoftwareRequirement$packages[[1]]$version
#> [1] "2.1.0"
#> 
#> 
#> 
#> 
#> 
#> $clt$`hisat2-mapping-se.cwl`$baseCommand
#> [1] "hisat2"
#> 
#> $clt$`hisat2-mapping-se.cwl`$requirements
#> $clt$`hisat2-mapping-se.cwl`$requirements$InitialWorkDirRequirement
#> $clt$`hisat2-mapping-se.cwl`$requirements$InitialWorkDirRequirement$listing
#> [1] "$(inputs.results_path)"
#> 
#> 
#> 
#> $clt$`hisat2-mapping-se.cwl`$arguments
#> $clt$`hisat2-mapping-se.cwl`$arguments[[1]]
#> $clt$`hisat2-mapping-se.cwl`$arguments[[1]]$prefix
#> [1] "-S"
#> 
#> $clt$`hisat2-mapping-se.cwl`$arguments[[1]]$valueFrom
#> [1] "$(inputs.results_path.path)/$(inputs.SampleName).sam"
#> 
#> 
#> $clt$`hisat2-mapping-se.cwl`$arguments[[2]]
#> $clt$`hisat2-mapping-se.cwl`$arguments[[2]]$prefix
#> [1] "-x"
#> 
#> $clt$`hisat2-mapping-se.cwl`$arguments[[2]]$valueFrom
#> [1] "$(inputs.idx_basedir.path)/$(inputs.idx_basename)"
#> 
#> 
#> $clt$`hisat2-mapping-se.cwl`$arguments[[3]]
#> $clt$`hisat2-mapping-se.cwl`$arguments[[3]]$prefix
#> [1] "-k"
#> 
#> $clt$`hisat2-mapping-se.cwl`$arguments[[3]]$valueFrom
#> [1] "1"
#> 
#> 
#> $clt$`hisat2-mapping-se.cwl`$arguments[[4]]
#> $clt$`hisat2-mapping-se.cwl`$arguments[[4]]$prefix
#> [1] "--min-intronlen"
#> 
#> $clt$`hisat2-mapping-se.cwl`$arguments[[4]]$valueFrom
#> [1] "30"
#> 
#> 
#> $clt$`hisat2-mapping-se.cwl`$arguments[[5]]
#> $clt$`hisat2-mapping-se.cwl`$arguments[[5]]$prefix
#> [1] "--max-intronlen"
#> 
#> $clt$`hisat2-mapping-se.cwl`$arguments[[5]]$valueFrom
#> [1] "3000"
#> 
#> 
#> 
#> $clt$`hisat2-mapping-se.cwl`$inputs
#> $clt$`hisat2-mapping-se.cwl`$inputs$idx_basedir
#> $clt$`hisat2-mapping-se.cwl`$inputs$idx_basedir$label
#> [1] "Path to the directory containing the index for the reference genome"
#> 
#> $clt$`hisat2-mapping-se.cwl`$inputs$idx_basedir$type
#> [1] "Directory"
#> 
#> 
#> $clt$`hisat2-mapping-se.cwl`$inputs$idx_basename
#> $clt$`hisat2-mapping-se.cwl`$inputs$idx_basename$label
#> [1] "Basename of the hisat2 index files"
#> 
#> $clt$`hisat2-mapping-se.cwl`$inputs$idx_basename$type
#> [1] "string"
#> 
#> 
#> $clt$`hisat2-mapping-se.cwl`$inputs$fq1
#> $clt$`hisat2-mapping-se.cwl`$inputs$fq1$label
#> [1] "Comma-separated list of files containing unpaired reads to be aligned"
#> 
#> $clt$`hisat2-mapping-se.cwl`$inputs$fq1$type
#> [1] "File"
#> 
#> $clt$`hisat2-mapping-se.cwl`$inputs$fq1$inputBinding
#> $clt$`hisat2-mapping-se.cwl`$inputs$fq1$inputBinding$prefix
#> [1] "-U"
#> 
#> $clt$`hisat2-mapping-se.cwl`$inputs$fq1$inputBinding$itemSeparator
#> [1] ","
#> 
#> 
#> 
#> $clt$`hisat2-mapping-se.cwl`$inputs$thread
#> $clt$`hisat2-mapping-se.cwl`$inputs$thread$label
#> [1] "Launch NTHREADS parallel search threads"
#> 
#> $clt$`hisat2-mapping-se.cwl`$inputs$thread$type
#> [1] "int"
#> 
#> $clt$`hisat2-mapping-se.cwl`$inputs$thread$inputBinding
#> $clt$`hisat2-mapping-se.cwl`$inputs$thread$inputBinding$prefix
#> [1] "--threads"
#> 
#> 
#> 
#> $clt$`hisat2-mapping-se.cwl`$inputs$SampleName
#> $clt$`hisat2-mapping-se.cwl`$inputs$SampleName$label
#> [1] "Filename to write output to"
#> 
#> $clt$`hisat2-mapping-se.cwl`$inputs$SampleName$type
#> [1] "string"
#> 
#> 
#> $clt$`hisat2-mapping-se.cwl`$inputs$results_path
#> $clt$`hisat2-mapping-se.cwl`$inputs$results_path$label
#> [1] "Path to the results directory"
#> 
#> $clt$`hisat2-mapping-se.cwl`$inputs$results_path$type
#> [1] "Directory"
#> 
#> 
#> 
#> $clt$`hisat2-mapping-se.cwl`$outputs
#> $clt$`hisat2-mapping-se.cwl`$outputs$hisat2_sam
#> $clt$`hisat2-mapping-se.cwl`$outputs$hisat2_sam$type
#> [1] "File"
#> 
#> $clt$`hisat2-mapping-se.cwl`$outputs$hisat2_sam$outputBinding
#> $clt$`hisat2-mapping-se.cwl`$outputs$hisat2_sam$outputBinding$glob
#> [1] "$(inputs.results_path.path)/$(inputs.SampleName).sam"
#> 
#> 
#> 
#> 
#> 
#> 
#> $yamlinput
#> $yamlinput$idx_basedir
#> $yamlinput$idx_basedir$class
#> [1] "Directory"
#> 
#> $yamlinput$idx_basedir$path
#> [1] "./data"
#> 
#> 
#> $yamlinput$idx_basename
#> [1] "tair10.fasta"
#> 
#> $yamlinput$fq1
#> $yamlinput$fq1$class
#> [1] "File"
#> 
#> $yamlinput$fq1$path
#> [1] "_FASTQ_PATH1_"
#> 
#> 
#> $yamlinput$SampleName
#> [1] "_SampleName_"
#> 
#> $yamlinput$results_path
#> $yamlinput$results_path$class
#> [1] "Directory"
#> 
#> $yamlinput$results_path$path
#> [1] "./results"
#> 
#> 
#> $yamlinput$thread
#> [1] 4
#> 
#> $yamlinput$ModulesToLoad
#> $yamlinput$ModulesToLoad$module1
#> [1] "hisat2/2.1.0"
#> 
#> 
#> 
#> $cmdlist
#> $cmdlist$M1A
#> $cmdlist$M1A$`hisat2-mapping-se`
#> [1] "hisat2 -S ./results/M1A.sam  -x ./data/tair10.fasta  -k 1  --min-intronlen 30  --max-intronlen 3000  -U ./data/SRR446027_1.fastq.gz --threads 4"
#> 
#> 
#> $cmdlist$M1B
#> $cmdlist$M1B$`hisat2-mapping-se`
#> [1] "hisat2 -S ./results/M1B.sam  -x ./data/tair10.fasta  -k 1  --min-intronlen 30  --max-intronlen 3000  -U ./data/SRR446028_1.fastq.gz --threads 4"
#> 
#> 
#> $cmdlist$A1A
#> $cmdlist$A1A$`hisat2-mapping-se`
#> [1] "hisat2 -S ./results/A1A.sam  -x ./data/tair10.fasta  -k 1  --min-intronlen 30  --max-intronlen 3000  -U ./data/SRR446029_1.fastq.gz --threads 4"
#> 
#> 
#> $cmdlist$A1B
#> $cmdlist$A1B$`hisat2-mapping-se`
#> [1] "hisat2 -S ./results/A1B.sam  -x ./data/tair10.fasta  -k 1  --min-intronlen 30  --max-intronlen 3000  -U ./data/SRR446030_1.fastq.gz --threads 4"
#> 
#> 
#> $cmdlist$V1A
#> $cmdlist$V1A$`hisat2-mapping-se`
#> [1] "hisat2 -S ./results/V1A.sam  -x ./data/tair10.fasta  -k 1  --min-intronlen 30  --max-intronlen 3000  -U ./data/SRR446031_1.fastq.gz --threads 4"
#> 
#> 
#> $cmdlist$V1B
#> $cmdlist$V1B$`hisat2-mapping-se`
#> [1] "hisat2 -S ./results/V1B.sam  -x ./data/tair10.fasta  -k 1  --min-intronlen 30  --max-intronlen 3000  -U ./data/SRR446032_1.fastq.gz --threads 4"
#> 
#> 
#> $cmdlist$M6A
#> $cmdlist$M6A$`hisat2-mapping-se`
#> [1] "hisat2 -S ./results/M6A.sam  -x ./data/tair10.fasta  -k 1  --min-intronlen 30  --max-intronlen 3000  -U ./data/SRR446033_1.fastq.gz --threads 4"
#> 
#> 
#> $cmdlist$M6B
#> $cmdlist$M6B$`hisat2-mapping-se`
#> [1] "hisat2 -S ./results/M6B.sam  -x ./data/tair10.fasta  -k 1  --min-intronlen 30  --max-intronlen 3000  -U ./data/SRR446034_1.fastq.gz --threads 4"
#> 
#> 
#> $cmdlist$A6A
#> $cmdlist$A6A$`hisat2-mapping-se`
#> [1] "hisat2 -S ./results/A6A.sam  -x ./data/tair10.fasta  -k 1  --min-intronlen 30  --max-intronlen 3000  -U ./data/SRR446035_1.fastq.gz --threads 4"
#> 
#> 
#> $cmdlist$A6B
#> $cmdlist$A6B$`hisat2-mapping-se`
#> [1] "hisat2 -S ./results/A6B.sam  -x ./data/tair10.fasta  -k 1  --min-intronlen 30  --max-intronlen 3000  -U ./data/SRR446036_1.fastq.gz --threads 4"
#> 
#> 
#> $cmdlist$V6A
#> $cmdlist$V6A$`hisat2-mapping-se`
#> [1] "hisat2 -S ./results/V6A.sam  -x ./data/tair10.fasta  -k 1  --min-intronlen 30  --max-intronlen 3000  -U ./data/SRR446037_1.fastq.gz --threads 4"
#> 
#> 
#> $cmdlist$V6B
#> $cmdlist$V6B$`hisat2-mapping-se`
#> [1] "hisat2 -S ./results/V6B.sam  -x ./data/tair10.fasta  -k 1  --min-intronlen 30  --max-intronlen 3000  -U ./data/SRR446038_1.fastq.gz --threads 4"
#> 
#> 
#> $cmdlist$M12A
#> $cmdlist$M12A$`hisat2-mapping-se`
#> [1] "hisat2 -S ./results/M12A.sam  -x ./data/tair10.fasta  -k 1  --min-intronlen 30  --max-intronlen 3000  -U ./data/SRR446039_1.fastq.gz --threads 4"
#> 
#> 
#> $cmdlist$M12B
#> $cmdlist$M12B$`hisat2-mapping-se`
#> [1] "hisat2 -S ./results/M12B.sam  -x ./data/tair10.fasta  -k 1  --min-intronlen 30  --max-intronlen 3000  -U ./data/SRR446040_1.fastq.gz --threads 4"
#> 
#> 
#> $cmdlist$A12A
#> $cmdlist$A12A$`hisat2-mapping-se`
#> [1] "hisat2 -S ./results/A12A.sam  -x ./data/tair10.fasta  -k 1  --min-intronlen 30  --max-intronlen 3000  -U ./data/SRR446041_1.fastq.gz --threads 4"
#> 
#> 
#> $cmdlist$A12B
#> $cmdlist$A12B$`hisat2-mapping-se`
#> [1] "hisat2 -S ./results/A12B.sam  -x ./data/tair10.fasta  -k 1  --min-intronlen 30  --max-intronlen 3000  -U ./data/SRR446042_1.fastq.gz --threads 4"
#> 
#> 
#> $cmdlist$V12A
#> $cmdlist$V12A$`hisat2-mapping-se`
#> [1] "hisat2 -S ./results/V12A.sam  -x ./data/tair10.fasta  -k 1  --min-intronlen 30  --max-intronlen 3000  -U ./data/SRR446043_1.fastq.gz --threads 4"
#> 
#> 
#> $cmdlist$V12B
#> $cmdlist$V12B$`hisat2-mapping-se`
#> [1] "hisat2 -S ./results/V12B.sam  -x ./data/tair10.fasta  -k 1  --min-intronlen 30  --max-intronlen 3000  -U ./data/SRR446044_1.fastq.gz --threads 4"
#> 
#> 
#> 
#> $input
#> $input$M1A
#> $input$M1A$FileName
#> [1] "./data/SRR446027_1.fastq.gz"
#> 
#> $input$M1A$SampleName
#> [1] "M1A"
#> 
#> 
#> $input$M1B
#> $input$M1B$FileName
#> [1] "./data/SRR446028_1.fastq.gz"
#> 
#> $input$M1B$SampleName
#> [1] "M1B"
#> 
#> 
#> $input$A1A
#> $input$A1A$FileName
#> [1] "./data/SRR446029_1.fastq.gz"
#> 
#> $input$A1A$SampleName
#> [1] "A1A"
#> 
#> 
#> $input$A1B
#> $input$A1B$FileName
#> [1] "./data/SRR446030_1.fastq.gz"
#> 
#> $input$A1B$SampleName
#> [1] "A1B"
#> 
#> 
#> $input$V1A
#> $input$V1A$FileName
#> [1] "./data/SRR446031_1.fastq.gz"
#> 
#> $input$V1A$SampleName
#> [1] "V1A"
#> 
#> 
#> $input$V1B
#> $input$V1B$FileName
#> [1] "./data/SRR446032_1.fastq.gz"
#> 
#> $input$V1B$SampleName
#> [1] "V1B"
#> 
#> 
#> $input$M6A
#> $input$M6A$FileName
#> [1] "./data/SRR446033_1.fastq.gz"
#> 
#> $input$M6A$SampleName
#> [1] "M6A"
#> 
#> 
#> $input$M6B
#> $input$M6B$FileName
#> [1] "./data/SRR446034_1.fastq.gz"
#> 
#> $input$M6B$SampleName
#> [1] "M6B"
#> 
#> 
#> $input$A6A
#> $input$A6A$FileName
#> [1] "./data/SRR446035_1.fastq.gz"
#> 
#> $input$A6A$SampleName
#> [1] "A6A"
#> 
#> 
#> $input$A6B
#> $input$A6B$FileName
#> [1] "./data/SRR446036_1.fastq.gz"
#> 
#> $input$A6B$SampleName
#> [1] "A6B"
#> 
#> 
#> $input$V6A
#> $input$V6A$FileName
#> [1] "./data/SRR446037_1.fastq.gz"
#> 
#> $input$V6A$SampleName
#> [1] "V6A"
#> 
#> 
#> $input$V6B
#> $input$V6B$FileName
#> [1] "./data/SRR446038_1.fastq.gz"
#> 
#> $input$V6B$SampleName
#> [1] "V6B"
#> 
#> 
#> $input$M12A
#> $input$M12A$FileName
#> [1] "./data/SRR446039_1.fastq.gz"
#> 
#> $input$M12A$SampleName
#> [1] "M12A"
#> 
#> 
#> $input$M12B
#> $input$M12B$FileName
#> [1] "./data/SRR446040_1.fastq.gz"
#> 
#> $input$M12B$SampleName
#> [1] "M12B"
#> 
#> 
#> $input$A12A
#> $input$A12A$FileName
#> [1] "./data/SRR446041_1.fastq.gz"
#> 
#> $input$A12A$SampleName
#> [1] "A12A"
#> 
#> 
#> $input$A12B
#> $input$A12B$FileName
#> [1] "./data/SRR446042_1.fastq.gz"
#> 
#> $input$A12B$SampleName
#> [1] "A12B"
#> 
#> 
#> $input$V12A
#> $input$V12A$FileName
#> [1] "./data/SRR446043_1.fastq.gz"
#> 
#> $input$V12A$SampleName
#> [1] "V12A"
#> 
#> 
#> $input$V12B
#> $input$V12B$FileName
#> [1] "./data/SRR446044_1.fastq.gz"
#> 
#> $input$V12B$SampleName
#> [1] "V12B"
#> 
#> 
#> 
#> $output
#> $output$M1A
#> $output$M1A$`hisat2-mapping-se`
#> [1] "./results/M1A.sam"
#> 
#> 
#> $output$M1B
#> $output$M1B$`hisat2-mapping-se`
#> [1] "./results/M1B.sam"
#> 
#> 
#> $output$A1A
#> $output$A1A$`hisat2-mapping-se`
#> [1] "./results/A1A.sam"
#> 
#> 
#> $output$A1B
#> $output$A1B$`hisat2-mapping-se`
#> [1] "./results/A1B.sam"
#> 
#> 
#> $output$V1A
#> $output$V1A$`hisat2-mapping-se`
#> [1] "./results/V1A.sam"
#> 
#> 
#> $output$V1B
#> $output$V1B$`hisat2-mapping-se`
#> [1] "./results/V1B.sam"
#> 
#> 
#> $output$M6A
#> $output$M6A$`hisat2-mapping-se`
#> [1] "./results/M6A.sam"
#> 
#> 
#> $output$M6B
#> $output$M6B$`hisat2-mapping-se`
#> [1] "./results/M6B.sam"
#> 
#> 
#> $output$A6A
#> $output$A6A$`hisat2-mapping-se`
#> [1] "./results/A6A.sam"
#> 
#> 
#> $output$A6B
#> $output$A6B$`hisat2-mapping-se`
#> [1] "./results/A6B.sam"
#> 
#> 
#> $output$V6A
#> $output$V6A$`hisat2-mapping-se`
#> [1] "./results/V6A.sam"
#> 
#> 
#> $output$V6B
#> $output$V6B$`hisat2-mapping-se`
#> [1] "./results/V6B.sam"
#> 
#> 
#> $output$M12A
#> $output$M12A$`hisat2-mapping-se`
#> [1] "./results/M12A.sam"
#> 
#> 
#> $output$M12B
#> $output$M12B$`hisat2-mapping-se`
#> [1] "./results/M12B.sam"
#> 
#> 
#> $output$A12A
#> $output$A12A$`hisat2-mapping-se`
#> [1] "./results/A12A.sam"
#> 
#> 
#> $output$A12B
#> $output$A12B$`hisat2-mapping-se`
#> [1] "./results/A12B.sam"
#> 
#> 
#> $output$V12A
#> $output$V12A$`hisat2-mapping-se`
#> [1] "./results/V12A.sam"
#> 
#> 
#> $output$V12B
#> $output$V12B$`hisat2-mapping-se`
#> [1] "./results/V12B.sam"
#> 
#> 
#> 
#> $cwlfiles
#> $cwlfiles$cwl
#> [1] "/tmp/RtmpOJ4T8t/temp_libpath3139e2165ead1/systemPipeR/extdata/cwl/hisat2/hisat2-se/hisat2-mapping-se.cwl"
#> 
#> $cwlfiles$yml
#> [1] "/tmp/RtmpOJ4T8t/temp_libpath3139e2165ead1/systemPipeR/extdata/cwl/hisat2/hisat2-se/hisat2-mapping-se.yml"
#> 
#> $cwlfiles$steps
#> [1] "hisat2-mapping-se"
#> 
#> $cwlfiles$targets
#> [1] "/tmp/RtmpOJ4T8t/temp_libpath3139e2165ead1/systemPipeR/extdata/targets.txt"
#> 
#> 
#> $inputvars
#> $inputvars$FileName
#> [1] "_FASTQ_PATH1_"
#> 
#> $inputvars$SampleName
#> [1] "_SampleName_"
#> 
#>