Class "SYSargsList"

S4 class container for storing parameters and information of command-line- or R-based workflow.

Objects from the Class

Objects can be created by calls of the form new("SYSargsList", ...).

Slots

sysconfig:

Object of class "list" storing all the config information for the project

codeSteps:

Object of class "list" storing all the code steps of the workflow

stepsWF:

Object of class "numeric" storing all the index steps levels in the workflow

dataWF:

Object of class "data.frame" storing all the data steps in the workflow

SYSargs2_steps:

Object of class "list" storing all the SYSargs2 objects of the workflow

statusWF:

Object of class "list" storing all the success and failure of each step in the workflow

projectWF:

Object of class "list" storing all the directories from the project

Methods

[

signature(x = "SYSargsList", i = "ANY", j = "ANY", drop = "ANY"): subsetting of class with bracket operator

[[

signature(x = "SYSargsList", i = "ANY", j = "ANY"): subsetting of class with bracket operator

[[<-

signature(x = "SYSargsList", i = "ANY", j = "ANY", value = "ANY"): replacement method for SYSargs2 class

$

signature(x = "SYSargsList"): extracting slots elements by name

codeSteps

signature(x = "SYSargsList"): ...

coerce

signature(from = "list", to = "SYSargsList"): as(list, "SYSargsList")

coerce

signature(from = "SYSargsList", to = "list"): as(SYSargsList, "list")

length

signature(x = "SYSargsList"): extracts number of SYSargsList objects

names

signature(x = "SYSargsList"): extracts slot names

projectWF

signature(x = "SYSargsList"): extract data from projectWF slot

show

signature(object = "SYSargsList"): summary view of SYSargsList steps

statusWF

signature(x = "SYSargsList"): extract data from statusWF slot

dataWF

signature(x = "SYSargsList"): extract data from dataWF slot

stepsWF

signature(x = "SYSargsList"): extract data from stepsWF slot

SYSargs2_steps

signature(x = "SYSargsList"): extract data from SYSargs2_steps slot

sysargslist

signature(x = "SYSargsList"): Coerce back to list as(SYSargsList, "list")

sysconfig

signature(x = "SYSargsList"): extract data from sysconfig slot

Author

Daniela Cassol and Thomas Girke

See also

Examples

showClass("SYSargsList")

#> Class "SYSargsList" [package "systemPipeR"]
#> 
#> Slots:
#>                                                                   
#> Name:       sysconfig      codeSteps        stepsWF         dataWF
#> Class:           list           list        numeric     data.frame
#>                                                    
#> Name:  SYSargs2_steps       statusWF      projectWF
#> Class:           list           list           list

if (FALSE) {
## Construct SYSargsList object from Rmd file
library(systemPipeRdata)
script <- system.file("extdata/workflows/rnaseq", "systemPipeRNAseq.Rmd", package="systemPipeRdata")
targets <- system.file("extdata", "targets.txt", package="systemPipeR")
dir_path <- tempdir()
SYSconfig <- initProject(projPath=dir_path, targets=targets, script=script, overwrite = TRUE)
sysargslist <- initWF(sysconfig ="SYSconfig.yml")
  }