Class "VENNset"

Container for storing Venn intersect results created by the overLapper function. The setlist slot stores the original label sets as vectors in a list; intersectmatrix organizes the label sets in a present-absent matrix; complexitylevels represents the number of comparisons considered for each comparison set as vector of integers; and vennlist contains the Venn intersect vectors.

Objects from the Class

Objects can be created by calls of the form new("VENNset", ...).

Slots

setlist:

Object of class "list": list of vectors

intersectmatrix:

Object of class "matrix": binary matrix

complexitylevels:

Object of class "integer": vector of integers

vennlist:

Object of class "list": list of vectors

Methods

as.list

signature(x = "VENNset"): coerces VENNset to list

coerce

signature(from = "list", to = "VENNset"): as(list, "VENNset")

complexitylevels

signature(x = "VENNset"): extracts data from complexitylevels slot

intersectmatrix

signature(x = "VENNset"): extracts data from intersectmatrix slot

length

signature(x = "VENNset"): returns number of original label sets

names

signature(x = "VENNset"): extracts slot names

setlist

signature(x = "VENNset"): extracts data from setlist slot

show

signature(object = "VENNset"): summary view of VENNset objects

vennlist

signature(x = "VENNset"): extracts data from vennset slot

Author

Thomas Girke

See also

Examples

showClass("VENNset")
#> Class "VENNset" [package "systemPipeR"]
#> 
#> Slots:
#>                                                                           
#> Name:           setlist  intersectmatrix complexitylevels         vennlist
#> Class:             list           matrix          integer             list

## Sample data
setlist <- list(A=sample(letters, 18), B=sample(letters, 16),
                C=sample(letters, 20), D=sample(letters, 22),
                E=sample(letters, 18), F=sample(letters, 22))

## Create VENNset
vennset <- overLapper(setlist[1:5], type="vennsets")
class(vennset)
#> [1] "VENNset"
#> attr(,"package")
#> [1] "systemPipeR"

## Accessor methods for VENNset/INTERSECTset objects
names(vennset)
#> [1] "setlist"          "intersectmatrix"  "complexitylevels" "vennlist"        
setlist(vennset)
#> $A
#>  [1] "m" "q" "h" "n" "f" "i" "l" "w" "t" "u" "j" "d" "y" "c" "b" "k" "a" "r"
#> 
#> $B
#>  [1] "s" "w" "y" "b" "o" "h" "g" "z" "e" "i" "v" "q" "u" "k" "a" "t"
#> 
#> $C
#>  [1] "t" "q" "a" "d" "z" "p" "v" "y" "c" "s" "u" "o" "b" "m" "k" "g" "e" "l" "x"
#> [20] "f"
#> 
#> $D
#>  [1] "b" "w" "a" "v" "k" "f" "x" "d" "n" "r" "c" "i" "u" "j" "q" "z" "o" "h" "p"
#> [20] "t" "l" "y"
#> 
#> $E
#>  [1] "g" "l" "z" "e" "a" "s" "o" "f" "t" "w" "k" "i" "b" "n" "x" "r" "h" "j"
#> 
intersectmatrix(vennset)
#>   A B C D E
#> a 1 1 1 1 1
#> b 1 1 1 1 1
#> c 1 0 1 1 0
#> d 1 0 1 1 0
#> e 0 1 1 0 1
#> f 1 0 1 1 1
#> g 0 1 1 0 1
#> h 1 1 0 1 1
#> i 1 1 0 1 1
#> j 1 0 0 1 1
#> k 1 1 1 1 1
#> l 1 0 1 1 1
#> m 1 0 1 0 0
#> n 1 0 0 1 1
#> o 0 1 1 1 1
#> p 0 0 1 1 0
#> q 1 1 1 1 0
#> r 1 0 0 1 1
#> s 0 1 1 0 1
#> t 1 1 1 1 1
#> u 1 1 1 1 0
#> v 0 1 1 1 0
#> w 1 1 0 1 1
#> x 0 0 1 1 1
#> y 1 1 1 1 0
#> z 0 1 1 1 1
complexitylevels(vennset)
#>  [1] 1 1 1 1 1 2 2 2 2 2 2 2 2 2 2 3 3 3 3 3 3 3 3 3 3 4 4 4 4 4 5
vennlist(vennset)
#> $A
#> character(0)
#> 
#> $B
#> character(0)
#> 
#> $C
#> character(0)
#> 
#> $D
#> character(0)
#> 
#> $E
#> character(0)
#> 
#> $A_B
#> character(0)
#> 
#> $A_C
#> [1] "m"
#> 
#> $A_D
#> character(0)
#> 
#> $A_E
#> character(0)
#> 
#> $B_C
#> character(0)
#> 
#> $B_D
#> character(0)
#> 
#> $B_E
#> character(0)
#> 
#> $C_D
#> [1] "p"
#> 
#> $C_E
#> character(0)
#> 
#> $D_E
#> character(0)
#> 
#> $A_B_C
#> character(0)
#> 
#> $A_B_D
#> character(0)
#> 
#> $A_B_E
#> character(0)
#> 
#> $A_C_D
#> [1] "c" "d"
#> 
#> $A_C_E
#> character(0)
#> 
#> $A_D_E
#> [1] "j" "n" "r"
#> 
#> $B_C_D
#> [1] "v"
#> 
#> $B_C_E
#> [1] "e" "g" "s"
#> 
#> $B_D_E
#> character(0)
#> 
#> $C_D_E
#> [1] "x"
#> 
#> $A_B_C_D
#> [1] "q" "u" "y"
#> 
#> $A_B_C_E
#> character(0)
#> 
#> $A_B_D_E
#> [1] "h" "i" "w"
#> 
#> $A_C_D_E
#> [1] "f" "l"
#> 
#> $B_C_D_E
#> [1] "o" "z"
#> 
#> $A_B_C_D_E
#> [1] "a" "b" "k" "t"
#> 

## Coerce VENNset/INTERSECTset object to list
as.list(vennset)
#> $setlist
#> $setlist$A
#>  [1] "m" "q" "h" "n" "f" "i" "l" "w" "t" "u" "j" "d" "y" "c" "b" "k" "a" "r"
#> 
#> $setlist$B
#>  [1] "s" "w" "y" "b" "o" "h" "g" "z" "e" "i" "v" "q" "u" "k" "a" "t"
#> 
#> $setlist$C
#>  [1] "t" "q" "a" "d" "z" "p" "v" "y" "c" "s" "u" "o" "b" "m" "k" "g" "e" "l" "x"
#> [20] "f"
#> 
#> $setlist$D
#>  [1] "b" "w" "a" "v" "k" "f" "x" "d" "n" "r" "c" "i" "u" "j" "q" "z" "o" "h" "p"
#> [20] "t" "l" "y"
#> 
#> $setlist$E
#>  [1] "g" "l" "z" "e" "a" "s" "o" "f" "t" "w" "k" "i" "b" "n" "x" "r" "h" "j"
#> 
#> 
#> $intersectmatrix
#>   A B C D E
#> a 1 1 1 1 1
#> b 1 1 1 1 1
#> c 1 0 1 1 0
#> d 1 0 1 1 0
#> e 0 1 1 0 1
#> f 1 0 1 1 1
#> g 0 1 1 0 1
#> h 1 1 0 1 1
#> i 1 1 0 1 1
#> j 1 0 0 1 1
#> k 1 1 1 1 1
#> l 1 0 1 1 1
#> m 1 0 1 0 0
#> n 1 0 0 1 1
#> o 0 1 1 1 1
#> p 0 0 1 1 0
#> q 1 1 1 1 0
#> r 1 0 0 1 1
#> s 0 1 1 0 1
#> t 1 1 1 1 1
#> u 1 1 1 1 0
#> v 0 1 1 1 0
#> w 1 1 0 1 1
#> x 0 0 1 1 1
#> y 1 1 1 1 0
#> z 0 1 1 1 1
#> 
#> $complexitylevels
#>  [1] 1 1 1 1 1 2 2 2 2 2 2 2 2 2 2 3 3 3 3 3 3 3 3 3 3 4 4 4 4 4 5
#> 
#> $vennlist
#> $vennlist$A
#> character(0)
#> 
#> $vennlist$B
#> character(0)
#> 
#> $vennlist$C
#> character(0)
#> 
#> $vennlist$D
#> character(0)
#> 
#> $vennlist$E
#> character(0)
#> 
#> $vennlist$A_B
#> character(0)
#> 
#> $vennlist$A_C
#> [1] "m"
#> 
#> $vennlist$A_D
#> character(0)
#> 
#> $vennlist$A_E
#> character(0)
#> 
#> $vennlist$B_C
#> character(0)
#> 
#> $vennlist$B_D
#> character(0)
#> 
#> $vennlist$B_E
#> character(0)
#> 
#> $vennlist$C_D
#> [1] "p"
#> 
#> $vennlist$C_E
#> character(0)
#> 
#> $vennlist$D_E
#> character(0)
#> 
#> $vennlist$A_B_C
#> character(0)
#> 
#> $vennlist$A_B_D
#> character(0)
#> 
#> $vennlist$A_B_E
#> character(0)
#> 
#> $vennlist$A_C_D
#> [1] "c" "d"
#> 
#> $vennlist$A_C_E
#> character(0)
#> 
#> $vennlist$A_D_E
#> [1] "j" "n" "r"
#> 
#> $vennlist$B_C_D
#> [1] "v"
#> 
#> $vennlist$B_C_E
#> [1] "e" "g" "s"
#> 
#> $vennlist$B_D_E
#> character(0)
#> 
#> $vennlist$C_D_E
#> [1] "x"
#> 
#> $vennlist$A_B_C_D
#> [1] "q" "u" "y"
#> 
#> $vennlist$A_B_C_E
#> character(0)
#> 
#> $vennlist$A_B_D_E
#> [1] "h" "i" "w"
#> 
#> $vennlist$A_C_D_E
#> [1] "f" "l"
#> 
#> $vennlist$B_C_D_E
#> [1] "o" "z"
#> 
#> $vennlist$A_B_C_D_E
#> [1] "a" "b" "k" "t"
#> 
#>