Convert targets to list or data.frame

Convert targets files to list or data.frame object.

targets.as.df(x)
targets.as.list(x, id="SampleName")

Arguments

x

An object of the class "list" or "data.frame" that stores data from each target file, as targets(WF).

id

A column from targets file, which will be used as an id for each one of the samples. It is required to be unique.

Value

data.frame or list containing all the targets file information.

See also

Author

Daniela Cassol

Examples

targetspath <- system.file("extdata", "targets.txt", package="systemPipeR")
targets <- read.delim(targetspath, comment.char = "#")
targetslist <- targets.as.list(x=targets)
targets.as.df(x=targetslist)
#>                       FileName SampleName Factor SampleLong Experiment
#> 1  ./data/SRR446027_1.fastq.gz        M1A     M1  Mock.1h.A          1
#> 2  ./data/SRR446028_1.fastq.gz        M1B     M1  Mock.1h.B          1
#> 3  ./data/SRR446029_1.fastq.gz        A1A     A1   Avr.1h.A          1
#> 4  ./data/SRR446030_1.fastq.gz        A1B     A1   Avr.1h.B          1
#> 5  ./data/SRR446031_1.fastq.gz        V1A     V1   Vir.1h.A          1
#> 6  ./data/SRR446032_1.fastq.gz        V1B     V1   Vir.1h.B          1
#> 7  ./data/SRR446033_1.fastq.gz        M6A     M6  Mock.6h.A          1
#> 8  ./data/SRR446034_1.fastq.gz        M6B     M6  Mock.6h.B          1
#> 9  ./data/SRR446035_1.fastq.gz        A6A     A6   Avr.6h.A          1
#> 10 ./data/SRR446036_1.fastq.gz        A6B     A6   Avr.6h.B          1
#> 11 ./data/SRR446037_1.fastq.gz        V6A     V6   Vir.6h.A          1
#> 12 ./data/SRR446038_1.fastq.gz        V6B     V6   Vir.6h.B          1
#> 13 ./data/SRR446039_1.fastq.gz       M12A    M12 Mock.12h.A          1
#> 14 ./data/SRR446040_1.fastq.gz       M12B    M12 Mock.12h.B          1
#> 15 ./data/SRR446041_1.fastq.gz       A12A    A12  Avr.12h.A          1
#> 16 ./data/SRR446042_1.fastq.gz       A12B    A12  Avr.12h.B          1
#> 17 ./data/SRR446043_1.fastq.gz       V12A    V12  Vir.12h.A          1
#> 18 ./data/SRR446044_1.fastq.gz       V12B    V12  Vir.12h.B          1
#>           Date
#> 1  23-Mar-2012
#> 2  23-Mar-2012
#> 3  23-Mar-2012
#> 4  23-Mar-2012
#> 5  23-Mar-2012
#> 6  23-Mar-2012
#> 7  23-Mar-2012
#> 8  23-Mar-2012
#> 9  23-Mar-2012
#> 10 23-Mar-2012
#> 11 23-Mar-2012
#> 12 23-Mar-2012
#> 13 23-Mar-2012
#> 14 23-Mar-2012
#> 15 23-Mar-2012
#> 16 23-Mar-2012
#> 17 23-Mar-2012
#> 18 23-Mar-2012