Generate targets file with reference

Generates targets file with sample-wise reference as required for some NGS applications, such as ChIP-Seq containing input samples. The reference sample information needs to be provided in the input file in a column called SampleReference where the values reference the labels used in the SampleName column. Sample rows without reference assignments will be removed automatically.

writeTargetsRef(infile, outfile, silent = FALSE, overwrite = FALSE, ...)

Arguments

infile
Path to input targets file.

outfile
Path to output targets file.

silent
If set to TRUE, all messages returned by the function will be suppressed.

overwrite
If set to TRUE, existing files of same name will be overwritten.

...
To pass on additional arguments.

Value

Generates modified targets file with the paths to the reference samples in the second column named FileName2. Note, sample rows not assigned reference samples are removed automatically.

Author

Thomas Girke

See also

Examples

## Path to input targets file
targets <- system.file("extdata", "targets_chip.txt", package="systemPipeR")

if (FALSE) {
## Write modified targets file with reference (e.g. input) sample
writeTargetsRef(infile=targets, outfile="~/targets_refsample.txt", silent=FALSE, overwrite=FALSE)
}