Writeout SYSargsList object

Function to writeout SYSargsList workflow control environment (S4 object) object.

write_SYSargsList(sysargs, sys.file=".SPRproject/SYSargsList.yml", silent=TRUE)

Arguments

sysargs
object of class SYSargsList.

sys.file
name and path of the SYSargsList file which will store all the project configuration information. Default is .SPRproject/SYSargsList.yml.

silent
if set to TRUE, all messages returned by the function will be suppressed.

Value

write_SYSargsList will return a sys.file path.

Author

Daniela Cassol

See also

Examples

## Construct SYSargsList object from Rmd file
sal <- SPRproject(overwrite=TRUE)

#> Recreating directory '/home/runner/work/systemPipeR/systemPipeR/docs/reference/.SPRproject'
#> Creating file '/home/runner/work/systemPipeR/systemPipeR/docs/reference/.SPRproject/SYSargsList.yml'

targetspath <- system.file("extdata/cwl/example/targets_example.txt", package="systemPipeR")

## Constructor and `appendStep<-`
appendStep(sal) <- SYSargsList(step_name = "echo", 
                               targets=targetspath, dir=TRUE,
                               wf_file="example/workflow_example.cwl", input_file="example/example.yml", 
                               dir_path = system.file("extdata/cwl", package="systemPipeR"),
                               inputvars = c(Message = "_STRING_", SampleName = "_SAMPLE_"))
appendStep(sal) <- LineWise(code = {
                            hello <- lapply(getColumn(sal, step=1, 'outfiles'), function(x) yaml::read_yaml(x))
                            }, 
                            step_name = "R_read", 
                            dependency = "echo")

sal <- write_SYSargsList(sal)
sal

#> [1] ".SPRproject/SYSargsList.yml"